refactor: reorganize token screen modules and components
