+++ src/ui/tokens/tokens_screen/mod.rs
@@ -1,4 +1,15 @@
-mod render_token_details;
+mod contract_details;
+mod data_contract_json_pop_up;
+mod distributions;
+mod groups;
+mod keyword_search;
+mod my_tokens;
+mod structs;
+mod token_creator;
+
+pub use structs::*;
+
+pub use groups::*;
 
 use std::collections::{BTreeMap, HashSet};
 use std::sync::atomic::Ordering;
@@ -28,12 +39,10 @@ use dash_sdk::dpp::data_contract::change_control_rules::ChangeControlRules;
 use dash_sdk::dpp::data_contract::conversion::json::DataContractJsonConversionMethodsV0;
 use dash_sdk::dpp::data_contract::group::v0::GroupV0;
 use dash_sdk::dpp::data_contract::group::{Group, GroupMemberPower, GroupRequiredPower};
-use dash_sdk::dpp::data_contract::{GroupContractPosition, TokenConfiguration, TokenContractPosition};
+use dash_sdk::dpp::data_contract::{GroupContractPosition, TokenConfiguration};
 use dash_sdk::dpp::data_contract::accessors::v0::DataContractV0Getters;
 use dash_sdk::dpp::data_contract::accessors::v1::DataContractV1Getters;
-use dash_sdk::dpp::data_contract::associated_token::token_perpetual_distribution::methods::v0::TokenPerpetualDistributionV0Accessors;
 use dash_sdk::dpp::fee::Credits;
-use dash_sdk::dpp::group::action_taker::{ActionGoal, ActionTaker};
 use dash_sdk::dpp::identity::accessors::IdentityGettersV0;
 use dash_sdk::dpp::identity::identity_public_key::accessors::v0::IdentityPublicKeyGettersV0;
 use dash_sdk::dpp::identity::{Purpose, SecurityLevel};
@@ -62,9 +71,7 @@ use crate::ui::components::left_panel::add_left_panel;
 use crate::ui::components::tokens_subscreen_chooser_panel::add_tokens_subscreen_chooser_panel;
 use crate::ui::components::top_panel::add_top_panel;
 use crate::ui::components::wallet_unlock::ScreenWithWalletUnlock;
-use crate::ui::{
-    BackendTaskSuccessResult, MessageType, RootScreenType, Screen, ScreenLike, ScreenType,
-};
+use crate::ui::{BackendTaskSuccessResult, MessageType, RootScreenType, ScreenLike, ScreenType};
 
 const EXP_FORMULA_PNG: &[u8] = include_bytes!("../../../../assets/exp_function.png");
 const INV_LOG_FORMULA_PNG: &[u8] = include_bytes!("../../../../assets/inv_log_function.png");
@@ -84,41 +91,6 @@ pub fn load_formula_image(bytes: &[u8]) -> ColorImage {
     let pixels = image.as_flat_samples();
     let color_image = ColorImage::from_rgba_unmultiplied(size, pixels.as_slice());
     color_image
-}
-
-/// Token info
-#[derive(Clone, Debug, PartialEq)]
-pub struct TokenInfo {
-    pub token_id: Identifier,
-    pub token_name: String,
-    pub data_contract_id: Identifier,
-    pub token_position: u16,
-    pub token_configuration: TokenConfiguration,
-    pub description: Option<String>,
-}
-
-#[derive(Clone, Debug, PartialEq)]
-pub struct TokenInfoWithDataContract {
-    pub token_id: Identifier,
-    pub token_name: String,
-    pub data_contract: DataContract,
-    pub token_position: u16,
-    pub token_configuration: TokenConfiguration,
-    pub description: Option<String>,
-}
-
-impl TokenInfoWithDataContract {
-    /// Constructs a `TokenInfoWithDataContract` from a `TokenInfo` and a `DataContract`.
-    pub fn from_with_data_contract(token_info: TokenInfo, data_contract: DataContract) -> Self {
-        TokenInfoWithDataContract {
-            token_id: token_info.token_id,
-            token_name: token_info.token_name,
-            data_contract,
-            token_position: token_info.token_position,
-            token_configuration: token_info.token_configuration,
-            description: token_info.description,
-        }
-    }
 }
 
 pub fn validate_perpetual_distribution_recipient(
@@ -190,109 +162,6 @@ fn sanitize_i64(input: &mut String) {
 
 fn sanitize_u64(input: &mut String) {
     input.retain(|c| c.is_ascii_digit());
-}
-
-#[derive(Clone, Copy, Debug, Eq, PartialEq, Hash, Ord, PartialOrd)]
-pub struct IdentityTokenIdentifier {
-    pub identity_id: Identifier,
-    pub token_id: Identifier,
-}
-
-impl From<IdentityTokenBalance> for IdentityTokenIdentifier {
-    fn from(value: IdentityTokenBalance) -> Self {
-        let IdentityTokenBalance {
-            token_id,
-            identity_id,
-            ..
-        } = value;
-
-        IdentityTokenIdentifier {
-            identity_id,
-            token_id,
-        }
-    }
-}
-
-impl From<IdentityTokenMaybeBalance> for IdentityTokenIdentifier {
-    fn from(value: IdentityTokenMaybeBalance) -> Self {
-        let IdentityTokenMaybeBalance {
-            token_id,
-            identity_id,
-            ..
-        } = value;
-
-        IdentityTokenIdentifier {
-            identity_id,
-            token_id,
-        }
-    }
-}
-
-/// A token owned by an identity.
-#[derive(Clone, Debug, PartialEq)]
-pub struct IdentityTokenMaybeBalance {
-    pub token_id: Identifier,
-    pub token_name: String,
-    pub identity_id: Identifier,
-    pub identity_alias: Option<String>,
-    pub balance: Option<IdentityTokenBalance>,
-}
-
-/// Identity Token Info
-#[derive(Clone, Debug, PartialEq)]
-pub struct IdentityTokenInfo {
-    pub token_id: Identifier,
-    pub token_alias: String,
-    pub identity: QualifiedIdentity,
-    pub data_contract: QualifiedContract,
-    pub token_config: TokenConfiguration,
-    pub token_position: TokenContractPosition,
-}
-
-impl IdentityTokenInfo {
-    pub fn try_from_identity_token_balance_with_lookup(
-        identity_token_balance: &IdentityTokenBalance,
-        app_context: &AppContext,
-    ) -> Result<Self, String> {
-        let IdentityTokenBalance {
-            token_id,
-            token_alias,
-            token_config,
-            identity_id,
-            data_contract_id,
-            token_position,
-            ..
-        } = identity_token_balance;
-        let identity = app_context
-            .get_identity_by_id(identity_id)
-            .map_err(|err| err.to_string())?
-            .ok_or("Identity not found".to_string())?;
-        let data_contract = app_context
-            .get_contract_by_id(data_contract_id)
-            .map_err(|err| err.to_string())?
-            .ok_or("Contract not found".to_string())?;
-        Ok(Self {
-            token_id: *token_id,
-            token_alias: token_alias.clone(),
-            identity,
-            data_contract,
-            token_config: token_config.clone(),
-            token_position: *token_position,
-        })
-    }
-}
-
-/// A token owned by an identity.
-#[derive(Clone, Debug, PartialEq)]
-pub struct IdentityTokenBalance {
-    pub token_id: Identifier,
-    pub token_alias: String,
-    pub token_config: TokenConfiguration,
-    pub identity_id: Identifier,
-    pub balance: TokenAmount,
-    pub estimated_unclaimed_rewards: Option<TokenAmount>,
-    pub data_contract_id: Identifier,
-    pub token_position: u16,
 }
 
 /// Which token sub-screen is currently showing.
@@ -1048,68 +917,6 @@ impl std::fmt::Display for TokenNameLanguage {
     }
 }
 
-#[derive(Default, Clone)]
-pub struct GroupMemberUI {
-    /// The base58 identity for this member
-    pub identity_str: String,
-    /// The power (u32) as a string for user input
-    pub power_str: String,
-}
-
-#[derive(Default, Clone)]
-pub struct GroupConfigUI {
-    /// Required power for the group (u32), user enters as string
-    pub required_power_str: String,
-    /// The members for this group
-    pub members: Vec<GroupMemberUI>,
-}
-
-impl GroupConfigUI {
-    /// Try converting this UI struct into a real `Group` (specifically `Group::V0`).
-    /// We also return the `u16` key that this group should be inserted under in the contract’s `groups` map.
-    fn parse_into_group(&self, pos: GroupContractPosition) -> Result<(u16, Group), String> {
-        // 1) Parse required power
-        let required_power = self.required_power_str.parse::<u32>().map_err(|_| {
-            format!(
-                "Invalid required power: '{}'. Must be an unsigned integer.",
-                self.required_power_str
-            )
-        })? as GroupRequiredPower;
-
-        // 2) Build a BTreeMap<Identifier, u32> for members
-        let mut members_map = BTreeMap::new();
-        for (i, member) in self.members.iter().enumerate() {
-            // A) Parse member identity from base58
-            let id =
-                Identifier::from_string(&member.identity_str, Encoding::Base58).map_err(|_| {
-                    format!(
-                        "Member #{}: invalid base58 identity '{}'",
-                        i + 1,
-                        member.identity_str
-                    )
-                })?;
-
-            // B) Parse power
-            let power =
-                member.power_str.parse::<u32>().map_err(|_| {
-                    format!("Member #{}: invalid power '{}'", i + 1, member.power_str)
-                })? as GroupMemberPower;
-
-            // Insert into the map
-            members_map.insert(id, power);
-        }
-
-        // 3) Construct Group::V0
-        let group_v0 = GroupV0 {
-            members: members_map,
-            required_power,
-        };
-
-        // 5) Return as (group_position, Group::V0 wrapped in Group::V0())
-        Ok((pos, Group::V0(group_v0)))
-    }
-}
-
 #[derive(Clone, Debug)]
 /// All arguments needed by `build_data_contract_v1_with_one_token`.
 pub struct TokenBuildArgs {
@@ -1622,33 +1429,6 @@ impl TokensScreen {
         screen
     }
 
-    pub fn change_to_preset(&mut self, preset: TokenConfigurationPreset) {
-        let basic_rules = preset.default_basic_change_control_rules_v0();
-        let advanced_rules = preset.default_advanced_change_control_rules_v0();
-        let emergency_rules = preset.default_emergency_action_change_control_rules_v0();
-
-        self.manual_minting_rules = basic_rules.clone().into();
-        self.manual_burning_rules = basic_rules.clone().into();
-        self.freeze_rules = advanced_rules.clone().into();
-        self.unfreeze_rules = advanced_rules.clone().into();
-        self.destroy_frozen_funds_rules = advanced_rules.clone().into();
-        self.emergency_action_rules = emergency_rules.clone().into();
-        self.max_supply_change_rules = advanced_rules.clone().into();
-        self.conventions_change_rules = basic_rules.clone().into();
-        self.perpetual_distribution_rules = advanced_rules.clone().into();
-        self.new_tokens_destination_identity_rules = basic_rules.clone().into();
-        self.minting_allow_choosing_destination_rules = basic_rules.clone().into();
-        self.authorized_main_control_group_change =
-            preset.default_main_control_group_can_be_modified();
-
-        // Reset optional identity/group inputs related to control group modification
-        self.main_control_group_change_authorized_identity = None;
-        self.main_control_group_change_authorized_group = None;
-
-        // Set `selected_token_preset` so UI shows current preset (Optional)
-        self.selected_token_preset = Some(preset.features);
-    }
-
     // ─────────────────────────────────────────────────────────────────
     // Reordering
     // ─────────────────────────────────────────────────────────────────
@@ -1722,329 +1502,6 @@ impl TokensScreen {
                 self.dismiss_message();
             }
         }
-    }
-
-    // ─────────────────────────────────────────────────────────────────
-    // Rendering
-    // ─────────────────────────────────────────────────────────────────
-
-    /// Renders the top-level token list (one row per unique token).
-    /// When the user clicks on a token, we set `selected_token_id`.
-    fn render_token_list(&mut self, ui: &mut Ui) -> Result<(), String> {
-        // Allocate space for refreshing indicator
-        let refreshing_height = 33.0;
-        let mut max_scroll_height = if let RefreshingStatus::Refreshing(_) = self.refreshing_status
-        {
-            ui.available_height() - refreshing_height
-        } else {
-            ui.available_height()
-        };
-
-        // Allocate space for backend message
-        let backend_message_height = 40.0;
-        if let Some((_, _, _)) = self.backend_message.clone() {
-            max_scroll_height -= backend_message_height;
-        }
-
-        // A simple table with columns: [Token Name | Token ID | Total Balance]
-        egui::ScrollArea::vertical()
-            .max_height(max_scroll_height)
-            .show(ui, |ui| {
-                Frame::group(ui.style())
-                    .fill(ui.visuals().panel_fill)
-                    .stroke(egui::Stroke::new(
-                        1.0,
-                        ui.visuals().widgets.inactive.bg_stroke.color,
-                    ))
-                    .inner_margin(Margin::same(8))
-                    .show(ui, |ui| {
-                        TableBuilder::new(ui)
-                            .striped(true)
-                            .resizable(true)
-                            .cell_layout(egui::Layout::left_to_right(Align::Center))
-                            .column(Column::initial(150.0).resizable(true)) // Token Name
-                            .column(Column::initial(200.0).resizable(true)) // Token ID
-                            .column(Column::initial(80.0).resizable(true)) // Description
-                            .column(Column::initial(80.0).resizable(true)) // Actions
-                            // .column(Column::initial(80.0).resizable(true)) // Token Info
-                            .header(30.0, |mut header| {
-                                header.col(|ui| {
-                                    ui.label("Token Name");
-                                });
-                                header.col(|ui| {
-                                    ui.label("Token ID");
-                                });
-                                header.col(|ui| {
-                                    ui.label("Description");
-                                });
-                                header.col(|ui| {
-                                    ui.label("Actions");
-                                });
-                            })
-                            .body(|mut body| {
-                                for TokenInfoWithDataContract {
-                                    token_id,
-                                    token_name,
-                                    description,
-                                    token_configuration,
-                                    data_contract,
-                                    ..
-                                } in self.all_known_tokens.values()
-                                {
-                                    body.row(25.0, |mut row| {
-                                        row.col(|ui| {
-                                            // By making the label into a button or using `ui.selectable_label`,
-                                            // we can respond to clicks.
-                                            if ui.button(token_name).clicked() {
-                                                self.selected_token = Some((
-                                                    token_id.clone(),
-                                                    data_contract.clone(),
-                                                    token_configuration.clone(),
-                                                ));
-                                            }
-                                        });
-                                        row.col(|ui| {
-                                            ui.label(token_id.to_string(Encoding::Base58));
-                                        });
-                                        row.col(|ui| {
-                                            ui.label(
-                                                description.as_ref().unwrap_or(&String::new()),
-                                            );
-                                        });
-                                        row.col(|ui| {
-                                            // Remove
-                                            if ui
-                                                .button("X")
-                                                .on_hover_text("Remove token from DET")
-                                                .clicked()
-                                            {
-                                                self.confirm_remove_token_popup = true;
-                                                self.token_to_remove = Some(token_id.clone());
-                                            }
-                                        });
-                                    });
-                                }
-                            });
-                    });
-            });
-        Ok(())
-    }
-
-    /// Renders details for the selected_contract_id.
-    fn render_contract_details(&mut self, ui: &mut Ui, contract_id: &Identifier) -> AppAction {
-        let mut action = AppAction::None;
-
-        if let Some(description) = &self.selected_contract_description {
-            ui.heading("Contract Description:");
-            ui.label(description.description.clone());
-        }
-
-        ui.add_space(10.0);
-
-        ui.heading("Tokens:");
-        let token_infos = self
-            .selected_token_infos
-            .iter()
-            .filter(|token| token.data_contract_id == *contract_id)
-            .cloned()
-            .collect::<Vec<_>>();
-        for token in token_infos {
-            if token.data_contract_id == *contract_id {
-                ui.heading(token.token_name.clone());
-                ui.label(format!(
-                    "ID: {}",
-                    token.token_id.to_string(Encoding::Base58)
-                ));
-                ui.label(format!(
-                    "Description: {}",
-                    token
-                        .description
-                        .clone()
-                        .unwrap_or("No description".to_string())
-                ));
-            }
-
-            ui.add_space(5.0);
-
-            // Add button to add token to my tokens
-            if ui.button("Add to My Tokens").clicked() {
-                match self.add_token_to_tracked_tokens(token.clone()) {
-                    Ok(internal_action) => {
-                        // Add token to my tokens
-                        action |= internal_action;
-                    }
-                    Err(e) => {
-                        self.set_error_message(Some(e));
-                    }
-                }
-            }
-
-            ui.add_space(10.0);
-        }
-
-        action
-    }
-
-    fn render_keyword_search(&mut self, ui: &mut Ui) -> AppAction {
-        let mut action = AppAction::None;
-
-        // 1) Input & “Go” button
-        ui.heading("Search Tokens by Keyword");
-        ui.add_space(5.0);
-
-        ui.horizontal(|ui| {
-            ui.label("Enter Keyword:");
-            let query_ref = self
-                .token_search_query
-                .get_or_insert_with(|| "".to_string());
-            let text_edit_response = ui.text_edit_singleline(query_ref);
-
-            let go_clicked = ui.button("Go").clicked();
-            let enter_pressed =
-                text_edit_response.lost_focus() && ui.input(|i| i.key_pressed(egui::Key::Enter));
-
-            if go_clicked || enter_pressed {
-                // Clear old results, set status
-                self.search_results.lock().unwrap().clear();
-                let now = Utc::now().timestamp() as u64;
-                self.contract_search_status = ContractSearchStatus::WaitingForResult(now);
-                self.search_current_page = 1;
-                self.next_cursors.clear();
-                self.previous_cursors.clear();
-                self.search_has_next_page = false;
-
-                // Dispatch a backend task to do the actual keyword => token retrieval
-                let keyword = query_ref.to_lowercase();
-                action = AppAction::BackendTask(BackendTask::TokenTask(
-                    TokenTask::QueryDescriptionsByKeyword(keyword, None),
-                ));
-            }
-        });
-
-        ui.add_space(10.0);
-
-        // 2) Display status
-        match &self.contract_search_status {
-            ContractSearchStatus::NotStarted => {
-                ui.label("Enter a keyword above and click Go.");
-            }
-            ContractSearchStatus::WaitingForResult(start_time) => {
-                let now = Utc::now().timestamp() as u64;
-                let elapsed = now - start_time;
-                ui.horizontal(|ui| {
-                    ui.label(format!("Searching... {} seconds", elapsed));
-                    ui.add(egui::widgets::Spinner::default().color(Color32::from_rgb(0, 128, 255)));
-                });
-            }
-            ContractSearchStatus::Complete => {
-                // Show the results
-                let results = self.search_results.lock().unwrap().clone();
-                if results.is_empty() {
-                    ui.label("No tokens match your keyword.");
-                } else {
-                    action |= self.render_search_results_table(ui, &results);
-                }
-
-                // Pagination controls
-                ui.horizontal(|ui| {
-                    if self.search_current_page > 1 {
-                        if ui.button("Previous").clicked() {
-                            // Go to previous page
-                            action = self.goto_previous_search_page();
-                        }
-                    }
-
-                    if !(self.next_cursors.is_empty() && self.previous_cursors.is_empty()) {
-                        ui.label(format!("Page {}", self.search_current_page));
-                    }
-
-                    if self.search_has_next_page {
-                        if ui.button("Next").clicked() {
-                            // Go to next page
-                            action = self.goto_next_search_page();
-                        }
-                    }
-                });
-            }
-            ContractSearchStatus::ErrorMessage(e) => {
-                ui.colored_label(Color32::RED, format!("Error: {}", e));
-            }
-        }
-
-        action
-    }
-
-    fn render_search_results_table(
-        &mut self,
-        ui: &mut Ui,
-        search_results: &[ContractDescriptionInfo],
-    ) -> AppAction {
-        let mut action = AppAction::None;
-
-        egui::ScrollArea::vertical().show(ui, |ui| {
-            Frame::group(ui.style())
-                .fill(ui.visuals().panel_fill)
-                .stroke(egui::Stroke::new(
-                    1.0,
-                    ui.visuals().widgets.inactive.bg_stroke.color,
-                ))
-                .inner_margin(Margin::same(8))
-                .show(ui, |ui| {
-                    TableBuilder::new(ui)
-                        .striped(true)
-                        .resizable(true)
-                        .cell_layout(egui::Layout::left_to_right(Align::Center))
-                        .column(Column::initial(60.0).resizable(true)) // Contract ID
-                        .column(Column::initial(200.0).resizable(true)) // Contract Description
-                        .column(Column::initial(80.0).resizable(true)) // Action
-                        .header(30.0, |mut header| {
-                            header.col(|ui| {
-                                ui.label("Contract ID");
-                            });
-                            header.col(|ui| {
-                                ui.label("Contract Description");
-                            });
-                            header.col(|ui| {
-                                ui.label("Action");
-                            });
-                        })
-                        .body(|mut body| {
-                            for contract in search_results {
-                                body.row(25.0, |mut row| {
-                                    row.col(|ui| {
-                                        ui.label(
-                                            contract.data_contract_id.to_string(Encoding::Base58),
-                                        );
-                                    });
-                                    row.col(|ui| {
-                                        ui.label(contract.description.clone());
-                                    });
-                                    row.col(|ui| {
-                                        // Example "Add" button
-                                        if ui.button("More Info").clicked() {
-                                            // Show more info about the token
-                                            self.selected_contract_id =
-                                                Some(contract.data_contract_id.clone());
-                                            action =
-                                                AppAction::BackendTask(BackendTask::ContractTask(
-                                                    ContractTask::FetchContractsWithDescriptions(
-                                                        vec![contract.data_contract_id.clone()],
-                                                    ),
-                                                ));
-
-                                            // // Add to MyTokens or do something with it
-                                            // // Note this is implemented, but we will add it back later!
-                                            // // We changed to searching contracts instead of tokens for now
-                                            // self.add_token_to_my_tokens(token.clone());
-                                        }
-                                    });
-                                });
-                            }
-                        });
-                });
-        });
-
-        action
     }
 
     fn history_row(&mut self, ui: &mut Ui) {
@@ -2156,1914 +1613,6 @@ impl TokensScreen {
         }
     }
 
-    pub fn render_token_creator(&mut self, context: &Context, ui: &mut Ui) -> AppAction {
-        let mut action = AppAction::None;
-
-        // 1) If we've successfully completed contract creation, show a success UI
-        if self.token_creator_status == TokenCreatorStatus::Complete {
-            self.render_token_creator_success_screen(ui);
-            return action;
-        }
-
-        // Allocate space for refreshing indicator
-        let refreshing_height = 33.0;
-        let mut max_scroll_height =
-            if let TokenCreatorStatus::WaitingForResult(_) = self.token_creator_status {
-                ui.available_height() - refreshing_height
-            } else {
-                ui.available_height()
-            };
-
-        // Allocate space for backend message
-        let backend_message_height = 40.0;
-        if let Some(_) = self.token_creator_error_message.clone() {
-            max_scroll_height -= backend_message_height;
-        }
-
-        egui::ScrollArea::vertical()
-            .max_height(max_scroll_height)
-            .show(ui, |ui| {
-                Frame::group(ui.style())
-                .fill(ui.visuals().panel_fill)
-                .stroke(egui::Stroke::new(
-                    1.0,
-                    ui.visuals().widgets.inactive.bg_stroke.color,
-                ))
-                .show(ui, |ui| {
-                    // Identity selection
-                    ui.add_space(10.0);
-                    let all_identities = match self.app_context.load_local_qualified_identities() {
-                        Ok(identities) => identities.into_iter().filter(|qi| !qi.private_keys.private_keys.is_empty()).collect::<Vec<_>>(),
-                        Err(_) => {
-                            ui.colored_label(Color32::RED, "Error loading identities from local DB");
-                            return;
-                        }
-                    };
-                    if all_identities.is_empty() {
-                        ui.colored_label(
-                                    Color32::DARK_RED,
-                                    "No identities loaded. Please load or create one to register the token contract with first.",
-                                );
-                        return;
-                    }
-
-                    ui.heading("1. Select an identity and key to register the token contract with:");
-                    ui.add_space(5.0);
-
-                    ui.horizontal(|ui| {
-                        ui.label("Identity:");
-                        ComboBox::from_id_salt("token_creator_identity_selector")
-                            .selected_text(
-                                self.selected_identity
-                                    .as_ref()
-                                    .map(|qi| {
-                                        qi.alias
-                                            .clone()
-                                            .unwrap_or_else(|| qi.identity.id().to_string(Encoding::Base58))
-                                    })
-                                    .unwrap_or_else(|| "Select Identity".to_owned()),
-                            )
-                            .show_ui(ui, |ui| {
-                                for identity in all_identities.iter() {
-                                    let display = identity
-                                        .alias
-                                        .clone()
-                                        .unwrap_or_else(|| identity.identity.id().to_string(Encoding::Base58));
-                                    if ui
-                                        .selectable_label(
-                                            Some(identity) == self.selected_identity.as_ref(),
-                                            display,
-                                        )
-                                        .clicked()
-                                    {
-                                        // On select, store it
-                                        self.selected_identity = Some(identity.clone());
-                                        // Clear the selected key & wallet
-                                        self.selected_key = None;
-                                        self.selected_wallet = None;
-                                        self.token_creator_error_message = None;
-                                    }
-                                }
-                            });
-                    });
-
-                    // Key selection
-                    ui.add_space(3.0);
-                    if let Some(ref qid) = self.selected_identity {
-                        // Attempt to list available keys (only auth keys in normal mode)
-                        let keys = if self.app_context.developer_mode.load(Ordering::Relaxed) {
-                            qid.identity
-                                .public_keys()
-                                .values()
-                                .cloned()
-                                .collect::<Vec<_>>()
-                        } else {
-                            qid.available_authentication_keys_with_critical_or_high_security_level()
-                                .into_iter()
-                                .map(|k| {
-                                    k.identity_public_key.clone()
-                                })
-                                .collect()
-                        };
-
-                        ui.horizontal(|ui| {
-                            ui.label("Key:");
-                            ComboBox::from_id_salt("token_creator_key_selector")
-                                .selected_text(match &self.selected_key {
-                                    Some(k) => format!(
-                                        "Key {} (Purpose: {:?}, Security Level: {:?})",
-                                        k.id(),
-                                        k.purpose(),
-                                        k.security_level()
-                                    ),
-                                    None => "Select Key".to_owned(),
-                                })
-                                .show_ui(ui, |ui| {
-                                    for key in keys {
-                                        let is_valid = key.purpose() == Purpose::AUTHENTICATION
-                                            && (key.security_level() == SecurityLevel::CRITICAL || key.security_level() == SecurityLevel::HIGH);
-
-                                        let label = format!(
-                                            "Key {} (Info: {}/{}/{})",
-                                            key.id(),
-                                            key.purpose(),
-                                            key.security_level(),
-                                            key.key_type()
-                                        );
-                                        let styled_label = if is_valid {
-                                            RichText::new(label.clone())
-                                        } else {
-                                            RichText::new(label.clone()).color(Color32::RED)
-                                        };
-
-                                        if ui
-                                            .selectable_label(
-                                                Some(key.id()) == self.selected_key.as_ref().map(|kk| kk.id()),
-                                                styled_label,
-                                            )
-                                            .clicked()
-                                        {
-                                            self.selected_key = Some(key.clone());
-
-                                            // If the key belongs to a wallet, set that wallet reference:
-                                            self.selected_wallet = crate::ui::identities::get_selected_wallet(
-                                                qid,
-                                                None,
-                                                Some(&key),
-                                                &mut self.token_creator_error_message,
-                                            );
-                                        }
-                                    }
-                                });
-                        });
-                    } else {
-                        ui.horizontal(|ui| {
-                            ui.label("Key:");
-                            ComboBox::from_id_salt("token_creator_key_selector_empty")
-                                .selected_text("Select Identity First")
-                                .show_ui(ui, |_| {
-                                });
-                        });
-                    }
-
-                    if self.selected_key.is_none() {
-                        return;
-                    }
-
-                    ui.add_space(10.0);
-                    ui.separator();
-
-                    // 3) If the wallet is locked, show unlock
-                    //    But only do this step if we actually have a wallet reference:
-                    let mut need_unlock = false;
-                    let mut just_unlocked = false;
-
-                    if let Some(_) = self.selected_wallet {
-                        let (n, j) = self.render_wallet_unlock_if_needed(ui);
-                        need_unlock = n;
-                        just_unlocked = j;
-                    }
-
-                    if need_unlock && !just_unlocked {
-                        // We must wait for unlock before continuing
-                        return;
-                    }
-
-                    // 4) Show input fields for token name, decimals, base supply, etc.
-                    ui.add_space(10.0);
-                    ui.heading("2. Enter basic token info:");
-                    ui.add_space(5.0);
-
-                    // Use `Grid` to align labels and text edits
-                    egui::Grid::new("basic_token_info_grid")
-                        .num_columns(2)
-                        .spacing([16.0, 8.0]) // Horizontal, vertical spacing
-                        .show(ui, |ui| {
-                            // Row 1: Token Name
-                            let mut token_to_remove: Option<u8> = None;
-                            for i in 0..self.token_names_input.len() {
-                                ui.label("Token Name (singular):");
-                                ui.text_edit_singleline(&mut self.token_names_input[i].0);
-                                if i == 0 {
-                                    ComboBox::from_id_salt(format!("token_name_language_selector_{}", i))
-                                        .selected_text(format!(
-                                            "{}",
-                                            self.token_names_input[i].2.to_string()
-                                        ))
-                                        .show_ui(ui, |ui| {
-                                            ui.selectable_value(&mut self.token_names_input[i].2, TokenNameLanguage::English, "English");
-                                        });
-                                } else {
-                                    ComboBox::from_id_salt(format!("token_name_language_selector_{}", i))
-                                        .selected_text(format!(
-                                            "{}",
-                                            self.token_names_input[i].2.to_string()
-                                        ))
-                                        .show_ui(ui, |ui| {
-                                            ui.selectable_value(&mut self.token_names_input[i].2, TokenNameLanguage::English, "English");
-                                            ui.selectable_value(&mut self.token_names_input[i].2, TokenNameLanguage::Arabic, "Arabic");
-                                            ui.selectable_value(&mut self.token_names_input[i].2, TokenNameLanguage::Bengali, "Bengali");
-                                            ui.selectable_value(&mut self.token_names_input[i].2, TokenNameLanguage::Burmese, "Burmese");
-                                            ui.selectable_value(&mut self.token_names_input[i].2, TokenNameLanguage::Chinese, "Chinese");
-                                            ui.selectable_value(&mut self.token_names_input[i].2, TokenNameLanguage::Czech, "Czech");
-                                            ui.selectable_value(&mut self.token_names_input[i].2, TokenNameLanguage::Dutch, "Dutch");
-                                            ui.selectable_value(&mut self.token_names_input[i].2, TokenNameLanguage::Farsi, "Farsi (Persian)");
-                                            ui.selectable_value(&mut self.token_names_input[i].2, TokenNameLanguage::Filipino, "Filipino (Tagalog)");
-                                            ui.selectable_value(&mut self.token_names_input[i].2, TokenNameLanguage::French, "French");
-                                            ui.selectable_value(&mut self.token_names_input[i].2, TokenNameLanguage::German, "German");
-                                            ui.selectable_value(&mut self.token_names_input[i].2, TokenNameLanguage::Greek, "Greek");
-                                            ui.selectable_value(&mut self.token_names_input[i].2, TokenNameLanguage::Gujarati, "Gujarati");
-                                            ui.selectable_value(&mut self.token_names_input[i].2, TokenNameLanguage::Hausa, "Hausa");
-                                            ui.selectable_value(&mut self.token_names_input[i].2, TokenNameLanguage::Hebrew, "Hebrew");
-                                            ui.selectable_value(&mut self.token_names_input[i].2, TokenNameLanguage::Hindi, "Hindi");
-                                            ui.selectable_value(&mut self.token_names_input[i].2, TokenNameLanguage::Hungarian, "Hungarian");
-                                            ui.selectable_value(&mut self.token_names_input[i].2, TokenNameLanguage::Igbo, "Igbo");
-                                            ui.selectable_value(&mut self.token_names_input[i].2, TokenNameLanguage::Indonesian, "Indonesian");
-                                            ui.selectable_value(&mut self.token_names_input[i].2, TokenNameLanguage::Italian, "Italian");
-                                            ui.selectable_value(&mut self.token_names_input[i].2, TokenNameLanguage::Japanese, "Japanese");
-                                            ui.selectable_value(&mut self.token_names_input[i].2, TokenNameLanguage::Javanese, "Javanese");
-                                            ui.selectable_value(&mut self.token_names_input[i].2, TokenNameLanguage::Kannada, "Kannada");
-                                            ui.selectable_value(&mut self.token_names_input[i].2, TokenNameLanguage::Khmer, "Khmer");
-                                            ui.selectable_value(&mut self.token_names_input[i].2, TokenNameLanguage::Korean, "Korean");
-                                            ui.selectable_value(&mut self.token_names_input[i].2, TokenNameLanguage::Malay, "Malay");
-                                            ui.selectable_value(&mut self.token_names_input[i].2, TokenNameLanguage::Malayalam, "Malayalam");
-                                            ui.selectable_value(&mut self.token_names_input[i].2, TokenNameLanguage::Mandarin, "Mandarin Chinese");
-                                            ui.selectable_value(&mut self.token_names_input[i].2, TokenNameLanguage::Marathi, "Marathi");
-                                            ui.selectable_value(&mut self.token_names_input[i].2, TokenNameLanguage::Nepali, "Nepali");
-                                            ui.selectable_value(&mut self.token_names_input[i].2, TokenNameLanguage::Oriya, "Oriya");
-                                            ui.selectable_value(&mut self.token_names_input[i].2, TokenNameLanguage::Pashto, "Pashto");
-                                            ui.selectable_value(&mut self.token_names_input[i].2, TokenNameLanguage::Polish, "Polish");
-                                            ui.selectable_value(&mut self.token_names_input[i].2, TokenNameLanguage::Portuguese, "Portuguese");
-                                            ui.selectable_value(&mut self.token_names_input[i].2, TokenNameLanguage::Punjabi, "Punjabi");
-                                            ui.selectable_value(&mut self.token_names_input[i].2, TokenNameLanguage::Romanian, "Romanian");
-                                            ui.selectable_value(&mut self.token_names_input[i].2, TokenNameLanguage::Russian, "Russian");
-                                            ui.selectable_value(&mut self.token_names_input[i].2, TokenNameLanguage::Serbian, "Serbian");
-                                            ui.selectable_value(&mut self.token_names_input[i].2, TokenNameLanguage::Sindhi, "Sindhi");
-                                            ui.selectable_value(&mut self.token_names_input[i].2, TokenNameLanguage::Sinhala, "Sinhala");
-                                            ui.selectable_value(&mut self.token_names_input[i].2, TokenNameLanguage::Somali, "Somali");
-                                            ui.selectable_value(&mut self.token_names_input[i].2, TokenNameLanguage::Spanish, "Spanish");
-                                            ui.selectable_value(&mut self.token_names_input[i].2, TokenNameLanguage::Swahili, "Swahili");
-                                            ui.selectable_value(&mut self.token_names_input[i].2, TokenNameLanguage::Swedish, "Swedish");
-                                            ui.selectable_value(&mut self.token_names_input[i].2, TokenNameLanguage::Tamil, "Tamil");
-                                            ui.selectable_value(&mut self.token_names_input[i].2, TokenNameLanguage::Telugu, "Telugu");
-                                            ui.selectable_value(&mut self.token_names_input[i].2, TokenNameLanguage::Thai, "Thai");
-                                            ui.selectable_value(&mut self.token_names_input[i].2, TokenNameLanguage::Turkish, "Turkish");
-                                            ui.selectable_value(&mut self.token_names_input[i].2, TokenNameLanguage::Ukrainian, "Ukrainian");
-                                            ui.selectable_value(&mut self.token_names_input[i].2, TokenNameLanguage::Urdu, "Urdu");
-                                            ui.selectable_value(&mut self.token_names_input[i].2, TokenNameLanguage::Vietnamese, "Vietnamese");
-                                            ui.selectable_value(&mut self.token_names_input[i].2, TokenNameLanguage::Yoruba, "Yoruba");
-                                            });
-                                }
-
-                                ui.horizontal(|ui| {
-                                    if ui.button("+").clicked() {
-                                        let used_languages: HashSet<_> = self.token_names_input.iter().map(|(_, _, lang, _)| *lang).collect();
-                                        let next_non_used_language = enum_iterator::all::<TokenNameLanguage>()
-                                            .find(|lang| !used_languages.contains(lang))
-                                            .unwrap_or(TokenNameLanguage::English);
-                                        // Add a new token name input
-                                        self.token_names_input.push((String::new(), String::new(), next_non_used_language, false));
-                                    }
-                                    if i != 0 {
-                                        if ui.button("-").clicked() {
-                                            token_to_remove = Some(i.try_into().expect("Failed to convert index"));
-                                        }
-                                    }
-
-                                    ui.checkbox(&mut self.token_names_input[i].3, "Add singular name to keywords");
-
-                                    let info_icon = Label::new("ℹ").sense(Sense::click());
-                                    let response = ui.add(info_icon)
-                                        .on_hover_text("Each searchable keyword costs 0.1 Dash");
-                                    if response.clicked() {
-                                        self.show_pop_up_info = Some("Each searchable keyword costs 0.1 Dash".to_string());
-                                    }
-                                    });
-                                ui.end_row();
-
-                                // Plural name
-                                ui.label("Token Name (plural):");
-                                ui.text_edit_singleline(&mut self.token_names_input[i].1);
-                                ui.end_row();
-
-                            }
-
-                            if let Some(token) = token_to_remove {
-                                self.token_names_input.remove(token.into());
-                            }
-
-                            // Row 2: Base Supply
-                            ui.label("Base Supply:");
-                            ui.text_edit_singleline(&mut self.base_supply_input);
-                            ui.end_row();
-
-                            // Row 3: Max Supply
-                            ui.label("Max Supply:");
-                            ui.text_edit_singleline(&mut self.max_supply_input);
-                            ui.end_row();
-
-                            // Row 4: Contract Keywords
-                            ui.horizontal(|ui| {
-                                ui.label("Contract Keywords (comma separated):");
-                                let info_icon = Label::new("ℹ").sense(Sense::click());
-                                let response = ui.add(info_icon)
-                                    .on_hover_text("Each searchable keyword costs 0.1 Dash");
-                                if response.clicked() {
-                                    self.show_pop_up_info = Some("Each searchable keyword costs 0.1 Dash".to_string());
-                                }
-                            });
-                            ui.text_edit_singleline(&mut self.contract_keywords_input);
-
-                            for name in self.token_names_input.iter() {
-                                if name.0.len() > 0 {
-                                    if name.3 {
-                                        let contract_keywords = self.contract_keywords_input.split(',').map(|s| s.trim()).collect::<Vec<_>>();
-
-                                        // If there are any duplicate keywords, show an error
-                                        let mut seen_keywords = HashSet::new();
-                                        seen_keywords.insert(name.0.clone());
-                                        for keyword in contract_keywords.iter() {
-                                            if seen_keywords.contains(*keyword) {
-                                                ui.colored_label(Color32::DARK_RED, format!("Duplicate contract keyword: {}", keyword));
-                                            }
-                                            seen_keywords.insert(keyword.to_string());
-                                        }
-                                    }
-                                }
-                            }
-                            ui.end_row();
-
-                            // Row 5: Token Description
-                            ui.label("Token Description (max 100 chars):");
-                            ui.text_edit_singleline(&mut self.token_description_input);
-                            ui.end_row();
-                        });
-
-                    ui.add_space(10.0);
-                    ui.separator();
-                    ui.add_space(10.0);
-
-                    // 5) Advanced settings toggle
-                    ui.collapsing("Advanced", |ui| {
-                        ui.add_space(3.0);
-
-                        // Use `Grid` to align labels and text edits
-                        egui::Grid::new("advanced_token_info_grid")
-                            .num_columns(2)
-                            .spacing([16.0, 8.0]) // Horizontal, vertical spacing
-                            .show(ui, |ui| {
-
-                                // Start as paused
-                                ui.horizontal(|ui| {
-                                    ui.checkbox(&mut self.start_as_paused_input, "Start as paused");
-
-                                    // Information icon with tooltip
-                                    if ui
-                                        .add(Label::new(RichText::new("ℹ").monospace()).sense(Sense::hover()))
-                                        .on_hover_text(
-                                            "When enabled, the token will be created in a paused state, meaning transfers will be \
-             disabled by default. All other token features—such as distributions and manual minting—\
-             remain fully functional. To allow transfers in the future, the token must be unpaused \
-             via an emergency action. It is strongly recommended to enable emergency actions if this \
-             option is selected, unless the intention is to permanently disable transfers.",
-                                        )
-                                        .hovered()
-                                    {
-                                        // Optional: visual feedback or styling if hovered
-                                    }
-                                });
-                                ui.end_row();
-
-                                self.history_row(ui);
-                                ui.end_row();
-
-                                // Name should be capitalized
-                                ui.horizontal(|ui| {
-                                    ui.checkbox(&mut self.should_capitalize_input, "Name should be capitalized");
-
-                                    // Information icon with tooltip
-                                    if ui
-                                        .add(Label::new(RichText::new("ℹ").monospace()).sense(Sense::hover()))
-                                        .on_hover_text(
-                                            "This is used only as helper information to client applications that will use \
-                                            token. This informs them on whether to capitalize the token name or not by default.",
-                                        )
-                                        .hovered()
-                                    {
-                                    }
-                                });
-                                ui.end_row();
-
-                                // Decimals
-                                ui.horizontal(|ui| {
-                                    ui.label("Max Decimals:");
-                                    // Restrict input to digits only
-                                    let response = ui.add(
-                                        TextEdit::singleline(&mut self.decimals_input).desired_width(50.0)
-                                    );
-
-                                    // Optionally filter out non-digit input
-                                    if response.changed() {
-                                        self.decimals_input.retain(|c| c.is_ascii_digit());
-                                        self.decimals_input.truncate(2);
-                                    }
-
-                                    let token_name = self.token_names_input
-                                        .first()
-                                        .as_ref()
-                                        .and_then(|(_, name, _, _)| if name.is_empty() { None} else { Some(name.as_str())})
-                                        .unwrap_or("<Token Name>");
-
-                                    let message = if self.decimals_input == "0" {
-                                        format!("Non Fractional Token (i.e 0, 1, 2 or 10 {})", token_name)
-                                    } else {
-                                        format!("Fractional Token (i.e 0.2 {})", token_name)
-                                    };
-
-                                    ui.label(RichText::new(message).color(Color32::GRAY));
-
-                                    if ui
-                                        .add(Label::new(RichText::new("ℹ").monospace()).sense(Sense::hover()))
-                                        .on_hover_text(
-                                            "The decimal places of the token, for example Dash and Bitcoin use 8. \
-                                            The minimum indivisible amount is a Duff or a Satoshi respectively. \
-                                            If you put a value greater than 0 this means that it is indicated that the \
-                                            consensus is that 10^(number entered) is what represents 1 full unit of the token.",
-                                        )
-                                        .hovered()
-                                    {
-                                    }
-                                });
-                                ui.end_row();
-                            });
-                    });
-
-                    ui.add_space(5.0);
-
-                    ui.collapsing("Action Rules", |ui| {
-                        ui.horizontal(|ui| {
-                            ui.label("Preset:");
-
-                            ComboBox::from_id_salt("preset_selector")
-                                .selected_text(
-                                    self.selected_token_preset
-                                        .map(|p|                                         match p {
-                                MostRestrictive => "Most Restrictive",
-                                WithOnlyEmergencyAction => "Only Emergency Action",
-                                WithMintingAndBurningActions => "Minting And Burning",
-                                WithAllAdvancedActions => "Advanced Actions",
-                                WithExtremeActions => "All Allowed",
-                            })
-                                        .unwrap_or("Custom"),
-                                )
-                                .show_ui(ui, |ui| {
-                                    use TokenConfigurationPresetFeatures::*;
-
-                                    // First, the "Custom" option
-                                    ui.selectable_value(
-                                        &mut self.selected_token_preset,
-                                        None,
-                                        "Custom",
-                                    );
-
-                                    for variant in [
-                                        MostRestrictive,
-                                        WithOnlyEmergencyAction,
-                                        WithMintingAndBurningActions,
-                                        WithAllAdvancedActions,
-                                        WithExtremeActions,
-                                    ] {
-                                        let text = match variant {
-                                            MostRestrictive => "Most Restrictive",
-                                            WithOnlyEmergencyAction => "Only Emergency Action",
-                                            WithMintingAndBurningActions => "Minting And Burning",
-                                            WithAllAdvancedActions => "Advanced Actions",
-                                            WithExtremeActions => "All Allowed",
-                                        };
-                                        if ui.selectable_value(
-                                            &mut self.selected_token_preset,
-                                            Some(variant),
-                                            text,
-                                        ).clicked() {
-                                            let preset = TokenConfigurationPreset {
-                                                features: variant,
-                                                action_taker: AuthorizedActionTakers::ContractOwner, // Or from a field the user selects
-                                            };
-                                            self.change_to_preset(preset);
-                                        }
-                                    }
-                                });
-                        });
-
-                        ui.add_space(3.0);
-
-                        self.manual_minting_rules.render_mint_control_change_rules_ui(ui, &self.groups_ui, &mut self.new_tokens_destination_identity_should_default_to_contract_owner, &mut self.new_tokens_destination_other_identity_enabled, &mut self.minting_allow_choosing_destination, &mut self.new_tokens_destination_identity_rules, &mut self.new_tokens_destination_other_identity, &mut self.minting_allow_choosing_destination_rules);
-                        self.manual_burning_rules.render_control_change_rules_ui(ui, &self.groups_ui,"Manual Burn", None);
-                        self.freeze_rules.render_control_change_rules_ui(ui, &self.groups_ui, "Freeze", Some(&mut self.allow_transfers_to_frozen_identities));
-                        self.unfreeze_rules.render_control_change_rules_ui(ui, &self.groups_ui, "Unfreeze", None);
-                        self.destroy_frozen_funds_rules.render_control_change_rules_ui(ui, &self.groups_ui, "Destroy Frozen Funds", None);
-                        self.emergency_action_rules.render_control_change_rules_ui(ui, &self.groups_ui, "Emergency Action", None);
-                        self.max_supply_change_rules.render_control_change_rules_ui(ui, &self.groups_ui, "Max Supply Change", None);
-                        self.conventions_change_rules.render_control_change_rules_ui(ui, &self.groups_ui, "Conventions Change", None);
-
-                        // Main control group change is slightly different so do this one manually.
-                        ui.collapsing("Main Control Group Change", |ui| {
-                            ui.add_space(3.0);
-
-                            // A) authorized_to_make_change
-                            ui.horizontal(|ui| {
-                                ui.label("Allow main control group change:");
-                                ComboBox::from_id_salt("main_control_group_change_selector")
-                                    .selected_text(format!(
-                                        "{}",
-                                        self.authorized_main_control_group_change.to_string()
-                                    ))
-                                    .show_ui(ui, |ui| {
-                                        ui.selectable_value(
-                                            &mut self.authorized_main_control_group_change,
-                                            AuthorizedActionTakers::NoOne,
-                                            "No One",
-                                        );
-                                        ui.selectable_value(
-                                            &mut self.authorized_main_control_group_change,
-                                            AuthorizedActionTakers::ContractOwner,
-                                            "Contract Owner",
-                                        );
-                                        ui.selectable_value(
-                                            &mut self.authorized_main_control_group_change,
-                                            AuthorizedActionTakers::Identity(Identifier::default()),
-                                            "Identity",
-                                        );
-                                        ui.selectable_value(
-                                            &mut self.authorized_main_control_group_change,
-                                            AuthorizedActionTakers::MainGroup,
-                                            "Main Group",
-                                        );
-                                        ui.selectable_value(
-                                            &mut self.authorized_main_control_group_change,
-                                            AuthorizedActionTakers::Group(0),
-                                            "Group",
-                                        );
-                                    });
-                                match &mut self.authorized_main_control_group_change {
-                                    AuthorizedActionTakers::Identity(_) => {
-                                        if self.main_control_group_change_authorized_identity.is_none() {
-                                            self.main_control_group_change_authorized_identity = Some(String::new());
-                                        }
-                                        if let Some(ref mut id) = self.main_control_group_change_authorized_identity {
-                                            ui.add(TextEdit::singleline(id).hint_text("base58 id"));
-                                        }
-                                    }
-                                    AuthorizedActionTakers::Group(_) => {
-                                        if self.main_control_group_change_authorized_group.is_none() {
-                                            self.main_control_group_change_authorized_group = Some("0".to_string());
-                                        }
-                                        if let Some(ref mut group) = self.main_control_group_change_authorized_group {
-                                            ui.add(TextEdit::singleline(group).hint_text("group contract position"));
-                                        }
-                                    }
-                                    _ => {}
-                                }
-                            });
-                        });
-                    });
-
-                    ui.add_space(5.0);
-
-                    ui.collapsing("Distribution", |ui| {
-                        ui.add_space(3.0);
-
-                        // PERPETUAL DISTRIBUTION SETTINGS
-                        if ui.checkbox(
-                            &mut self.enable_perpetual_distribution,
-                            "Enable Perpetual Distribution",
-                        ).clicked() {
-                            self.perpetual_dist_type = PerpetualDistributionIntervalTypeUI::TimeBased;
-                        };
-                        if self.enable_perpetual_distribution {
-                            ui.add_space(5.0);
-
-                            // 2) Select the distribution type
-                            ui.horizontal(|ui| {
-                                ui.label("     Type:");
-                                ComboBox::from_id_salt("perpetual_dist_type_selector")
-                                    .selected_text(format!("{:?}", self.perpetual_dist_type))
-                                    .show_ui(ui, |ui| {
-                                        ui.selectable_value(
-                                            &mut self.perpetual_dist_type,
-                                            PerpetualDistributionIntervalTypeUI::TimeBased,
-                                            "Time-Based",
-                                        );
-                                        ui.selectable_value(
-                                            &mut self.perpetual_dist_type,
-                                            PerpetualDistributionIntervalTypeUI::EpochBased,
-                                            "Epoch-Based",
-                                        );
-                                        ui.selectable_value(
-                                            &mut self.perpetual_dist_type,
-                                            PerpetualDistributionIntervalTypeUI::BlockBased,
-                                            "Block-Based",
-                                        );
-                                    });
-                            });
-
-                            // If user picked a real distribution type:
-                            match self.perpetual_dist_type {
-                                PerpetualDistributionIntervalTypeUI::TimeBased => {
-                                    ui.add_space(2.0);
-                                    ui.horizontal(|ui| {
-                                        ui.label("        - Distributes every ");
-
-                                        // Restrict input to digits only
-                                        let response = ui.add(
-                                            TextEdit::singleline(&mut self.perpetual_dist_interval_input)
-                                        );
-
-                                        // Optionally filter out non-digit input
-                                        if response.changed() {
-                                            self.perpetual_dist_interval_input.retain(|c| c.is_ascii_digit());
-                                        }
-
-                                        // Dropdown for selecting unit
-                                        ComboBox::from_id_salt("interval_unit_selector")
-                                            .selected_text(
-                                                self.perpetual_dist_interval_unit
-                                                    .label_for_amount(&self.perpetual_dist_interval_input),
-                                            )
-                                            .show_ui(ui, |ui| {
-                                                for unit in [
-                                                    IntervalTimeUnit::Second,
-                                                    IntervalTimeUnit::Minute,
-                                                    IntervalTimeUnit::Hour,
-                                                    IntervalTimeUnit::Day,
-                                                    IntervalTimeUnit::Week,
-                                                    IntervalTimeUnit::Year,
-                                                ] {
-                                                    ui.selectable_value(
-                                                        &mut self.perpetual_dist_interval_unit,
-                                                        unit.clone(),
-                                                        unit.label_for_amount(&self.perpetual_dist_interval_input),
-                                                    );
-                                                }
-                                            });
-                                    });
-                                }
-                                PerpetualDistributionIntervalTypeUI::EpochBased => {
-                                    ui.add_space(2.0);
-                                    ui.horizontal(|ui| {
-                                        ui.label("        - Distributes every (Epoch Interval):");
-                                        ui.text_edit_singleline(&mut self.perpetual_dist_interval_input);
-                                    });
-                                }
-                                PerpetualDistributionIntervalTypeUI::BlockBased => {
-                                    ui.add_space(2.0);
-                                    ui.horizontal(|ui| {
-                                        ui.label("        - Distributes every (Block Interval):");
-                                        ui.text_edit_singleline(&mut self.perpetual_dist_interval_input);
-                                    });
-                                }
-                                PerpetualDistributionIntervalTypeUI::None => {
-                                    // Do nothing
-                                }
-                            }
-
-                            ui.add_space(10.0);
-
-                            // 3) Select the distribution function
-                            ui.horizontal(|ui| {
-                                ui.label("     Function:");
-                                ComboBox::from_id_salt("perpetual_dist_function_selector")
-                                    .selected_text(format!("{:?}", self.perpetual_dist_function))
-                                    .show_ui(ui, |ui| {
-                                        ui.selectable_value(
-                                            &mut self.perpetual_dist_function,
-                                            DistributionFunctionUI::FixedAmount,
-                                            "FixedAmount",
-                                        );
-                                        ui.selectable_value(
-                                            &mut self.perpetual_dist_function,
-                                            DistributionFunctionUI::StepDecreasingAmount,
-                                            "StepDecreasing",
-                                        );
-                                        ui.selectable_value(
-                                            &mut self.perpetual_dist_function,
-                                            DistributionFunctionUI::Stepwise,
-                                            "Stepwise",
-                                        );
-                                        ui.selectable_value(
-                                            &mut self.perpetual_dist_function,
-                                            DistributionFunctionUI::Linear,
-                                            "Linear",
-                                        );
-                                        ui.selectable_value(
-                                            &mut self.perpetual_dist_function,
-                                            DistributionFunctionUI::Polynomial,
-                                            "Polynomial",
-                                        );
-                                        ui.selectable_value(
-                                            &mut self.perpetual_dist_function,
-                                            DistributionFunctionUI::Exponential,
-                                            "Exponential",
-                                        );
-                                        ui.selectable_value(
-                                            &mut self.perpetual_dist_function,
-                                            DistributionFunctionUI::Logarithmic,
-                                            "Logarithmic",
-                                        );
-                                        ui.selectable_value(
-                                            &mut self.perpetual_dist_function,
-                                            DistributionFunctionUI::InvertedLogarithmic,
-                                            "InvertedLogarithmic",
-                                        );
-                                    });
-
-                                    let info_icon = Label::new("ℹ").sense(Sense::click());
-                                    let response = ui.add(info_icon).on_hover_text("Info about distribution types");
-
-                                    // Check if the label was clicked
-                                    if response.clicked() {
-                                        self.show_pop_up_info = Some(r#"
-# FixedAmount
-
-Emits a constant (fixed) number of tokens for every period.
-
-### Formula
-For any period `x`, the emitted tokens are:
-
-`f(x) = n`
-
-### Use Case
-- When a predictable, unchanging reward is desired.
-- Simplicity and stable emissions.
-
-### Example
-- If `n = 5` tokens per block, then after 3 blocks the total emission is 15 tokens.
-
----
-
-# StepDecreasingAmount
-
-Emits a random number of tokens within a specified range.
-
-### Description
-- This function selects a **random** token emission amount between `min` and `max`.
-- The value is drawn **uniformly** between the bounds.
-- The randomness uses a Pseudo Random Function (PRF) from x.
-
-### Formula
-For any period `x`, the emitted tokens follow:
-
-`f(x) ∈ [min, max]`
-
-### Parameters
-- `min`: The **minimum** possible number of tokens emitted.
-- `max`: The **maximum** possible number of tokens emitted.
-
-### Use Cases
-- **Stochastic Rewards**: Introduces randomness into rewards to incentivize unpredictability.
-- **Lottery-Based Systems**: Used for randomized emissions, such as block rewards with probabilistic payouts.
-
-### Example
-Suppose a system emits **between 10 and 100 tokens per period**.
-
-`Random { min: 10, max: 100 }`
-
-| Period (x) | Emitted Tokens (Random) |
-|------------|------------------------|
-| 1          | 27                     |
-| 2          | 94                     |
-| 3          | 63                     |
-| 4          | 12                     |
-
-- Each period, the function emits a **random number of tokens** between `min = 10` and `max = 100`.
-- Over time, the **average reward trends toward the midpoint** `(min + max) / 2`.
-
-### Constraints
-- **`min` must be ≤ `max`**, otherwise the function is invalid.
-- If `min == max`, this behaves like a `FixedAmount` function with a constant emission.
-
----
-
-### LinearInteger
-
-A linear function using integer precision.
-
-- **Formula:** f(x) = a * x + b
-- **Description:**
-    - a > 0 -> tokens increase over time
-    - a < 0 -> tokens decrease over time
-    - b is the initial value
-- **Use Case:** Incentivize early or match ecosystem growth
-- **Example:** f(x) = 10x + 50
-
----
-
-### LinearFloat
-
-A linear function with fractional (floating-point) rates.
-
-- **Formula:** f(x) = a * x + b
-- **Description:** Similar to LinearInteger, but with fractional slope
-- **Use Case:** Gradual fractional increases/decreases over time
-- **Example:** f(x) = 0.5x + 50
-
----
-
-### PolynomialInteger
-
-A polynomial function (e.g. quadratic, cubic) using integer precision.
-
-- **Formula:** f(x) = a * x^n + b
-- **Description:** Flexible curves (growth/decay) beyond simple linear.
-- **Use Case:** Diminishing or accelerating returns as time progresses
-- **Example:** f(x) = 2x^2 + 20
-
----
-
-### PolynomialFloat
-
-A polynomial function supporting fractional exponents or coefficients.
-
-- **Formula:** f(x) = a * x^n + b
-- **Description:** Similar to PolynomialInteger, but with floats
-- **Example:** f(x) = 0.5x^3 + 20
-
----
-
-### Exponential
-
-Exponential growth or decay of tokens.
-
-- **Formula:** f(x) = a * e^(b * x) + c
-- **Description:**
-    - b > 0 -> rapid growth
-    - b < 0 -> rapid decay
-- **Use Case:** Early contributor boosts or quick emission tapering
-- **Example:** f(x) = 100 * e^(-0.693 * x) + 5
-
----
-
-### Logarithmic
-
-Logarithmic growth of token emissions.
-
-- **Formula:** f(x) = a * log_b(x) + c
-- **Description:** Growth slows as x increases.
-- **Use Case:** Sustainable long-term emission tapering
-- **Example:** f(x) = 20 * log_2(x) + 5
-
----
-
-### Stepwise
-
-Emits tokens in fixed amounts for specific intervals.
-
-- **Description:** Emissions remain constant within each step.
-- **Use Case:** Adjust rewards at specific milestones
-- **Example:** 100 tokens per block for first 1000 blocks, then 50 tokens thereafter.
-"#
-                                        .to_string())};
-                            });
-
-                            ui.add_space(2.0);
-
-                            if let Some(texture) = self.function_textures.get(&self.perpetual_dist_function) {
-                                ui.add_space(10.0);
-                                ui.horizontal(|ui| {
-                                    ui.add_space(50.0); // Shift image right
-                                    ui.image(texture);
-                                });
-                                ui.add_space(10.0);
-                            } else if let Some(image) = self.function_images.get(&self.perpetual_dist_function) {
-                                let texture = context.load_texture(self.perpetual_dist_function.name(), image.clone(), Default::default());
-                                self.function_textures.insert(self.perpetual_dist_function.clone(), texture.clone());
-                                ui.add_space(10.0);
-                                ui.horizontal(|ui| {
-                                    ui.add_space(50.0); // Shift image right
-                                    ui.image(&texture);
-                                });
-                                ui.add_space(10.0);
-                            }
-
-                            // Based on the user’s chosen function, display relevant fields:
-                            match self.perpetual_dist_function {
-                                DistributionFunctionUI::FixedAmount => {
-                                    ui.horizontal(|ui| {
-                                        ui.label("        - Fixed Amount per Interval:");
-                                        ui.text_edit_singleline(&mut self.fixed_amount_input);
-                                    });
-                                }
-
-                                DistributionFunctionUI::StepDecreasingAmount => {
-                                    ui.horizontal(|ui| {
-                                        ui.label("        - Step Count:");
-                                        let response = ui.add(TextEdit::singleline(&mut self.step_count_input));
-                                        if response.changed() {
-                                            sanitize_u64(&mut self.step_count_input);
-                                        }
-                                        if !self.step_count_input.is_empty() {
-                                            if let Ok((perpetual_dist_interval_input, step_count_input)) = self.perpetual_dist_interval_input.parse::<u64>().and_then(|perpetual_dist_interval_input| self.step_count_input.parse::<u64>().map(|step_count_input| (perpetual_dist_interval_input, step_count_input))) {
-                                                let text = match self.perpetual_dist_type {
-                                                    PerpetualDistributionIntervalTypeUI::None => "".to_string(),
-                                                    PerpetualDistributionIntervalTypeUI::BlockBased => {
-                                                        let amount = perpetual_dist_interval_input * step_count_input;
-                                                        if amount == 1 {
-                                                            "Every Block".to_string()
-                                                        } else {
-                                                            format!("Every {} Blocks", amount)
-                                                        }
-                                                    }
-                                                    PerpetualDistributionIntervalTypeUI::TimeBased => {
-                                                        let amount = perpetual_dist_interval_input * step_count_input;
-                                                        format!("Every {} {}", amount, self.perpetual_dist_interval_unit.capitalized_label_for_num_amount(amount))
-                                                    }
-                                                    PerpetualDistributionIntervalTypeUI::EpochBased => {
-                                                        let amount = perpetual_dist_interval_input * step_count_input;
-                                                        if amount == 1 {
-                                                            "Every Epoch Change".to_string()
-                                                        } else {
-                                                            format!("Every {} Epochs", amount)
-                                                        }
-                                                    }
-                                                };
-
-                                                ui.label(RichText::new(text).color(Color32::GRAY));
-                                            }
-                                        }
-                                    });
-
-                                    ui.horizontal(|ui| {
-                                        ui.label("        - Decrease per Interval Numerator (n < 65,536):");
-                                        let response = ui.add(TextEdit::singleline(&mut self.decrease_per_interval_numerator_input));
-                                        if response.changed() {
-                                            sanitize_u64(&mut self.decrease_per_interval_numerator_input);
-                                            self.decrease_per_interval_numerator_input.truncate(5);
-                                        }
-                                    });
-
-                                    ui.horizontal(|ui| {
-                                        ui.label("        - Decrease per Interval Denominator (d < 65,536):");
-                                        let response = ui.add(TextEdit::singleline(&mut self.decrease_per_interval_denominator_input));
-                                        if response.changed() {
-                                            sanitize_u64(&mut self.decrease_per_interval_denominator_input);
-                                            self.decrease_per_interval_denominator_input.truncate(5);
-                                        }
-                                    });
-
-                                    ui.horizontal(|ui| {
-                                        ui.label("        - Start Period Offset (i64, optional):");
-                                        let response = ui.add(
-                                            TextEdit::singleline(&mut self.step_decreasing_start_period_offset_input)
-                                                .hint_text("None"),
-                                        );
-                                        if response.changed() {
-                                            sanitize_i64(&mut self.step_decreasing_start_period_offset_input);
-                                        }
-                                    });
-
-                                    ui.horizontal(|ui| {
-                                        ui.label("        - Initial Token Emission Amount:");
-                                        let response = ui.add(TextEdit::singleline(&mut self.step_decreasing_initial_emission_input));
-                                        if response.changed() {
-                                            sanitize_u64(&mut self.step_decreasing_initial_emission_input);
-                                        }
-                                    });
-
-                                    ui.horizontal(|ui| {
-                                        ui.label("        - Minimum Emission Value (optional):");
-                                        let response = ui.add(
-                                            TextEdit::singleline(&mut self.step_decreasing_min_value_input)
-                                                .hint_text("None"),
-                                        );
-                                        if response.changed() {
-                                            sanitize_u64(&mut self.step_decreasing_min_value_input);
-                                        }
-                                    });
-
-                                    ui.horizontal(|ui| {
-                                        ui.label("        - Maximum Interval Count (optional):");
-                                        let response = ui.add(
-                                            TextEdit::singleline(&mut self.step_decreasing_max_interval_count_input)
-                                                .hint_text("None"),
-                                        );
-                                        if response.changed() {
-                                            sanitize_u64(&mut self.step_decreasing_max_interval_count_input);
-                                        }
-                                    });
-
-                                    ui.horizontal(|ui| {
-                                        ui.label("        - Trailing Distribution Interval Token Amount:");
-                                        let response = ui.add(
-                                            TextEdit::singleline(&mut self.step_decreasing_trailing_distribution_interval_amount_input),
-                                        );
-                                        if response.changed() {
-                                            sanitize_u64(&mut self.step_decreasing_trailing_distribution_interval_amount_input);
-                                        }
-                                    });
-                                }
-
-                                DistributionFunctionUI::Stepwise => {
-                                    // Example: multiple steps (u64 block => some token amount).
-                                    // Each element in `stepwise_steps` is (String, String) = (block, amount).
-                                    // You can show them in a loop and let users edit each pair.
-                                    let mut i = 0;
-                                    while i < self.stepwise_steps.len() {
-                                        let (mut steps_str, mut amount_str) = self.stepwise_steps[i].clone();
-
-                                        ui.horizontal(|ui| {
-                                            ui.label(format!("        - Step #{}:", i));
-                                            ui.label("Start Step:");
-                                            let response = ui.add(TextEdit::singleline(&mut steps_str).desired_width(50.0));
-                                            if response.changed() {
-                                                sanitize_u64(&mut steps_str);
-                                            }
-
-                                            ui.label("Amount:");
-                                            let response = ui.add(TextEdit::singleline(&mut amount_str).desired_width(50.0));
-                                            if response.changed() {
-                                                sanitize_u64(&mut amount_str);
-                                            }
-
-                                            if let Ok((perpetual_dist_interval_input, step_position)) = self.perpetual_dist_interval_input.parse::<u64>().and_then(|perpetual_dist_interval_input| steps_str.parse::<u64>().map(|step_count_input| (perpetual_dist_interval_input, step_count_input))) {
-                                                if let Ok(amount) = amount_str.parse::<u64>() {
-                                                    let every_text = match self.perpetual_dist_type {
-                                                        PerpetualDistributionIntervalTypeUI::None => "".to_string(),
-                                                        PerpetualDistributionIntervalTypeUI::BlockBased => {
-                                                            if perpetual_dist_interval_input == 1 {
-                                                                "every block".to_string()
-                                                            } else {
-                                                                format!("every {} blocks", perpetual_dist_interval_input)
-                                                            }
-                                                        }
-                                                        PerpetualDistributionIntervalTypeUI::TimeBased => {
-                                                            format!("every {} {}", perpetual_dist_interval_input, self.perpetual_dist_interval_unit.label_for_num_amount(perpetual_dist_interval_input))
-                                                        }
-                                                        PerpetualDistributionIntervalTypeUI::EpochBased => {
-                                                            if perpetual_dist_interval_input == 1 {
-                                                                "every epoch change".to_string()
-                                                            } else {
-                                                                format!("every {} epochs", perpetual_dist_interval_input)
-                                                            }
-                                                        }
-                                                    };
-
-                                                    let text = match self.perpetual_dist_type {
-                                                        PerpetualDistributionIntervalTypeUI::None => "".to_string(),
-                                                        PerpetualDistributionIntervalTypeUI::BlockBased => {
-                                                            let block = step_position * perpetual_dist_interval_input;
-                                                            if block == 0 {
-                                                                if amount == 0 {
-                                                                    "At start don't distribute tokens".to_string()
-                                                                } else {
-                                                                    format!("At start distribute {} tokens {}", amount, every_text)
-                                                                }
-                                                            } else {
-                                                                if amount == 0 {
-                                                                    format!("After block {} stop distributing tokens", block)
-                                                                } else {
-                                                                    format!("After block {} distribute {} tokens {}", block, amount, every_text)
-                                                                }
-                                                            }
-                                                        }
-                                                        PerpetualDistributionIntervalTypeUI::TimeBased => {
-                                                            let time = step_position * perpetual_dist_interval_input;
-                                                            if time == 0 {
-                                                                if amount == 0 {
-                                                                    "At start don't distribute tokens".to_string()
-                                                                } else {
-                                                                    format!("At start distribute {} tokens {}", amount, every_text)
-                                                                }
-                                                            } else {
-                                                                if amount == 0 {
-                                                                    format!("{} {} after the contract is registered stop distributing tokens", time, self.perpetual_dist_interval_unit.label_for_num_amount(perpetual_dist_interval_input))
-                                                                } else {
-                                                                    format!("{} {} after the contract is registered distribute {} tokens {}", time, self.perpetual_dist_interval_unit.label_for_num_amount(perpetual_dist_interval_input), amount, every_text)
-                                                                }
-                                                            }
-                                                        }
-                                                        PerpetualDistributionIntervalTypeUI::EpochBased => {
-                                                            let epoch = step_position * perpetual_dist_interval_input;
-                                                            if epoch == 0 {
-                                                                if amount == 0 {
-                                                                    "At start don't distribute tokens".to_string()
-                                                                } else {
-                                                                    format!("At start distribute {} tokens {}", amount, every_text)
-                                                                }
-                                                            } else {
-                                                                if amount == 0 {
-                                                                    format!("After epoch {} stop distributing tokens", epoch)
-                                                                } else {
-                                                                    format!("After epoch {} distribute {} tokens {}", epoch, amount, every_text)
-                                                                }
-                                                            }
-                                                        }
-                                                    };
-
-                                                    ui.label(RichText::new(text).color(Color32::GRAY));
-                                                }
-
-
-                                            }
-
-                                            // If remove is clicked, remove the step at index i
-                                            // and *do not* increment i, because the next element
-                                            // now “shifts” into this index.
-                                            if ui.button("Remove").clicked() {
-                                                self.stepwise_steps.remove(i);
-                                            } else {
-                                                // Otherwise, update the vector with any edits and move to the next step
-                                                self.stepwise_steps[i] = (steps_str, amount_str);
-                                                i += 1;
-                                            }
-                                        });
-                                    }
-
-                                    // A button to add new steps
-                                    ui.horizontal(|ui| {
-                                        ui.label("     ");
-                                        if ui.button("Add Step").clicked() {
-                                            self.stepwise_steps.push(("0".to_owned(), "0".to_owned()));
-                                        }
-                                    });
-                                }
-
-                                DistributionFunctionUI::Linear => {
-                                    ui.horizontal(|ui| {
-                                        ui.label("        - Slope Numerator (a, { -255 ≤ a ≤ 256 }):");
-                                        let response = ui.add(TextEdit::singleline(&mut self.linear_int_a_input));
-                                        if response.changed() {
-                                            sanitize_i64(&mut self.linear_int_a_input);
-                                        }
-                                    });
-                                    ui.horizontal(|ui| {
-                                        ui.label("        - Slope Divisor (d, u64):");
-                                        let response = ui.add(TextEdit::singleline(&mut self.linear_int_d_input));
-                                        if response.changed() {
-                                            sanitize_u64(&mut self.linear_int_d_input);
-                                        }
-                                    });
-                                    ui.horizontal(|ui| {
-                                        ui.label("        - Start Step (s, i64, optional):");
-                                        let response = ui.add(
-                                            TextEdit::singleline(&mut self.linear_int_start_step_input)
-                                                .hint_text("None"),
-                                        );
-                                        if response.changed() {
-                                            sanitize_i64(&mut self.linear_int_start_step_input);
-                                        }
-                                    });
-                                    ui.horizontal(|ui| {
-                                        ui.label("        - Starting Amount (b, i64):");
-                                        let response = ui.add(TextEdit::singleline(&mut self.linear_int_starting_amount_input));
-                                        if response.changed() {
-                                            sanitize_i64(&mut self.linear_int_starting_amount_input);
-                                        }
-                                    });
-                                    ui.horizontal(|ui| {
-                                        ui.label("        - Minimum Emission Value (optional):");
-                                        let response = ui.add(
-                                            TextEdit::singleline(&mut self.linear_int_min_value_input)
-                                                .hint_text("None"),
-                                        );
-                                        if response.changed() {
-                                            sanitize_u64(&mut self.linear_int_min_value_input);
-                                        }
-                                    });
-                                    ui.horizontal(|ui| {
-                                        ui.label("        - Maximum Emission Value (optional):");
-                                        let response = ui.add(
-                                            TextEdit::singleline(&mut self.linear_int_max_value_input)
-                                                .hint_text("None"),
-                                        );
-                                        if response.changed() {
-                                            sanitize_u64(&mut self.linear_int_max_value_input);
-                                        }
-                                    });
-                                }
-
-                                DistributionFunctionUI::Polynomial => {
-                                    ui.horizontal(|ui| {
-                                        ui.label("        - Scaling Factor (a, { -255 ≤ a ≤ 256 }):");
-                                        ui.text_edit_singleline(&mut self.poly_int_a_input);
-                                        sanitize_i64(&mut self.poly_int_a_input);
-                                    });
-
-                                    ui.horizontal(|ui| {
-                                        ui.label("        - Exponent Numerator (m, { -8 ≤ m ≤ 8 }):");
-                                        ui.text_edit_singleline(&mut self.poly_int_m_input);
-                                        sanitize_i64(&mut self.poly_int_m_input);
-                                    });
-
-                                    ui.horizontal(|ui| {
-                                        ui.label("        - Exponent Denominator (n, { 0 < n ≤ 32 }):");
-                                        ui.text_edit_singleline(&mut self.poly_int_n_input);
-                                        sanitize_u64(&mut self.poly_int_n_input);
-                                    });
-
-                                    ui.horizontal(|ui| {
-                                        ui.label("        - Divisor (d, u64):");
-                                        ui.text_edit_singleline(&mut self.poly_int_d_input);
-                                        sanitize_u64(&mut self.poly_int_d_input);
-                                    });
-
-                                    ui.horizontal(|ui| {
-                                        ui.label("        - Start Period Offset (s, optional, u64):");
-                                        let response = ui.add(
-                                            TextEdit::singleline(&mut self.poly_int_s_input)
-                                                .hint_text("None"),
-                                        );
-                                        if response.changed() && !self.poly_int_s_input.trim().is_empty() {
-                                            sanitize_u64(&mut self.poly_int_s_input);
-                                        }
-                                    });
-
-                                    ui.horizontal(|ui| {
-                                        ui.label("        - Offset (o, i64):");
-                                        ui.text_edit_singleline(&mut self.poly_int_o_input);
-                                        sanitize_i64(&mut self.poly_int_o_input);
-                                    });
-
-                                    ui.horizontal(|ui| {
-                                        ui.label("        - Initial Token Emission (b, u64):");
-                                        ui.text_edit_singleline(&mut self.poly_int_b_input);
-                                        sanitize_u64(&mut self.poly_int_b_input);
-                                    });
-
-                                    ui.horizontal(|ui| {
-                                        ui.label("        - Minimum Emission Value (optional, u64):");
-                                        let response = ui.add(
-                                            TextEdit::singleline(&mut self.poly_int_min_value_input)
-                                                .hint_text("None"),
-                                        );
-                                        if response.changed() && !self.poly_int_min_value_input.trim().is_empty() {
-                                            sanitize_u64(&mut self.poly_int_min_value_input);
-                                        }
-                                    });
-
-                                    ui.horizontal(|ui| {
-                                        ui.label("        - Maximum Emission Value (optional, u64):");
-                                        let response = ui.add(
-                                            TextEdit::singleline(&mut self.poly_int_max_value_input)
-                                                .hint_text("None"),
-                                        );
-                                        if response.changed() && !self.poly_int_max_value_input.trim().is_empty() {
-                                            sanitize_u64(&mut self.poly_int_max_value_input);
-                                        }
-                                    });
-                                }
-
-                                DistributionFunctionUI::Exponential => {
-                                    ui.horizontal(|ui| {
-                                        ui.label("        - Scaling Factor (a, { 0 < a ≤ 256 }):");
-                                        ui.text_edit_singleline(&mut self.exp_a_input);
-                                        sanitize_u64(&mut self.exp_a_input);
-                                    });
-                                    ui.horizontal(|ui| {
-                                        ui.label("        - Exponent Rate Numerator (m, { -8 ≤ m ≤ 8 ; m ≠ 0 }):");
-                                        ui.text_edit_singleline(&mut self.exp_m_input);
-                                        sanitize_i64(&mut self.exp_m_input);
-                                    });
-                                    ui.horizontal(|ui| {
-                                        ui.label("        - Exponent Rate Denominator (n, { 0 < n ≤ 32 }):");
-                                        ui.text_edit_singleline(&mut self.exp_n_input);
-                                        sanitize_u64(&mut self.exp_n_input);
-                                    });
-                                    ui.horizontal(|ui| {
-                                        ui.label("        - Divisor (d, { u64 ; d ≠ 0 }):");
-                                        ui.text_edit_singleline(&mut self.exp_d_input);
-                                        sanitize_u64(&mut self.exp_d_input);
-                                    });
-                                    ui.horizontal(|ui| {
-                                        ui.label("        - Start Period Offset (s, optional, u64):");
-                                        let response = ui.add(
-                                            TextEdit::singleline(&mut self.exp_s_input)
-                                                .hint_text("None"),
-                                        );
-                                        if response.changed() && !self.exp_s_input.trim().is_empty() {
-                                            sanitize_u64(&mut self.exp_s_input);
-                                        }
-                                    });
-                                    ui.horizontal(|ui| {
-                                        ui.label("        - Offset (o, i64):");
-                                        ui.text_edit_singleline(&mut self.exp_o_input);
-                                        sanitize_i64(&mut self.exp_o_input);
-                                    });
-                                    ui.horizontal(|ui| {
-                                        ui.label("        - Base amount (b, u64):");
-                                        ui.text_edit_singleline(&mut self.exp_b_input);
-                                        sanitize_i64(&mut self.exp_b_input);
-                                    });
-                                    ui.horizontal(|ui| {
-                                        ui.label("        - Minimum Emission Value (optional, u64):");
-                                        let response = ui.add(
-                                            TextEdit::singleline(&mut self.exp_min_value_input)
-                                                .hint_text("None"),
-                                        );
-                                        if response.changed() && !self.exp_min_value_input.trim().is_empty() {
-                                            sanitize_u64(&mut self.exp_min_value_input);
-                                        }
-                                    });
-
-                                    ui.horizontal(|ui| {
-                                        ui.label("        - Maximum Emission Value (optional, u64):");
-                                        let response = ui.add(
-                                            TextEdit::singleline(&mut self.exp_max_value_input)
-                                                .hint_text("None"),
-                                        );
-                                        if response.changed() && !self.exp_max_value_input.trim().is_empty() {
-                                            sanitize_u64(&mut self.exp_max_value_input);
-                                        }
-                                    });
-                                }
-
-                                DistributionFunctionUI::Logarithmic => {
-                                    ui.horizontal(|ui| {
-                                        ui.label("        - Scaling Factor (a, i64, { -32_766 ≤ a ≤ 32_767 }):");
-                                        ui.text_edit_singleline(&mut self.log_a_input);
-                                        sanitize_i64(&mut self.log_a_input);
-                                    });
-
-                                    ui.horizontal(|ui| {
-                                        ui.label("        - Divisor (d, u64):");
-                                        ui.text_edit_singleline(&mut self.log_d_input);
-                                        sanitize_u64(&mut self.log_d_input);
-                                    });
-
-                                    ui.horizontal(|ui| {
-                                        ui.label("        - Exponent Numerator (m, u64):");
-                                        ui.text_edit_singleline(&mut self.log_m_input);
-                                        sanitize_u64(&mut self.log_m_input);
-                                    });
-
-                                    ui.horizontal(|ui| {
-                                        ui.label("        - Exponent Denominator (n, u64):");
-                                        ui.text_edit_singleline(&mut self.log_n_input);
-                                        sanitize_u64(&mut self.log_n_input);
-                                    });
-
-                                    ui.horizontal(|ui| {
-                                        ui.label("        - Start Period Offset (s, optional, u64):");
-                                        let response = ui.add(
-                                            TextEdit::singleline(&mut self.log_s_input)
-                                                .hint_text("None"),
-                                        );
-                                        if response.changed() && !self.log_s_input.trim().is_empty() {
-                                            sanitize_u64(&mut self.log_s_input);
-                                        }
-                                    });
-
-                                    ui.horizontal(|ui| {
-                                        ui.label("        - Offset (o, i64):");
-                                        ui.text_edit_singleline(&mut self.log_o_input);
-                                        sanitize_i64(&mut self.log_o_input);
-                                    });
-
-                                    ui.horizontal(|ui| {
-                                        ui.label("        - Base Amount (b, u64):");
-                                        ui.text_edit_singleline(&mut self.log_b_input);
-                                        sanitize_u64(&mut self.log_b_input);
-                                    });
-
-                                    ui.horizontal(|ui| {
-                                        ui.label("        - Minimum Emission Value (optional, u64):");
-                                        let response = ui.add(
-                                            TextEdit::singleline(&mut self.log_min_value_input)
-                                                .hint_text("None"),
-                                        );
-                                        if response.changed() && !self.log_min_value_input.trim().is_empty() {
-                                            sanitize_u64(&mut self.log_min_value_input);
-                                        }
-                                    });
-
-                                    ui.horizontal(|ui| {
-                                        ui.label("        - Maximum Emission Value (optional, u64):");
-                                        let response = ui.add(
-                                            TextEdit::singleline(&mut self.log_max_value_input)
-                                                .hint_text("None"),
-                                        );
-                                        if response.changed() && !self.log_max_value_input.trim().is_empty() {
-                                            sanitize_u64(&mut self.log_max_value_input);
-                                        }
-                                    });
-                                }
-
-                                DistributionFunctionUI::InvertedLogarithmic => {
-                                    ui.horizontal(|ui| {
-                                        ui.label("        - Scaling Factor (a, i64, { -32_766 ≤ a ≤ 32_767 }):");
-                                        ui.text_edit_singleline(&mut self.inv_log_a_input);
-                                        sanitize_i64(&mut self.inv_log_a_input);
-                                    });
-
-                                    ui.horizontal(|ui| {
-                                        ui.label("        - Divisor (d, u64):");
-                                        ui.text_edit_singleline(&mut self.inv_log_d_input);
-                                        sanitize_u64(&mut self.inv_log_d_input);
-                                    });
-
-                                    ui.horizontal(|ui| {
-                                        ui.label("        - Exponent Numerator (m, u64):");
-                                        ui.text_edit_singleline(&mut self.inv_log_m_input);
-                                        sanitize_u64(&mut self.inv_log_m_input);
-                                    });
-
-                                    ui.horizontal(|ui| {
-                                        ui.label("        - Exponent Denominator (n, u64):");
-                                        ui.text_edit_singleline(&mut self.inv_log_n_input);
-                                        sanitize_u64(&mut self.inv_log_n_input);
-                                    });
-
-                                    ui.horizontal(|ui| {
-                                        ui.label("        - Start Period Offset (s, optional, u64):");
-                                        let response = ui.add(
-                                            TextEdit::singleline(&mut self.inv_log_s_input)
-                                                .hint_text("None"),
-                                        );
-                                        if response.changed() && !self.inv_log_s_input.trim().is_empty() {
-                                            sanitize_u64(&mut self.inv_log_s_input);
-                                        }
-                                    });
-
-                                    ui.horizontal(|ui| {
-                                        ui.label("        - Offset (o, i64):");
-                                        ui.text_edit_singleline(&mut self.inv_log_o_input);
-                                        sanitize_i64(&mut self.inv_log_o_input);
-                                    });
-
-                                    ui.horizontal(|ui| {
-                                        ui.label("        - Base Amount (b, u64):");
-                                        ui.text_edit_singleline(&mut self.inv_log_b_input);
-                                        sanitize_u64(&mut self.inv_log_b_input);
-                                    });
-
-                                    ui.horizontal(|ui| {
-                                        ui.label("        - Minimum Emission Value (optional, u64):");
-                                        let response = ui.add(
-                                            TextEdit::singleline(&mut self.inv_log_min_value_input)
-                                                .hint_text("None"),
-                                        );
-                                        if response.changed() && !self.inv_log_min_value_input.trim().is_empty() {
-                                            sanitize_u64(&mut self.inv_log_min_value_input);
-                                        }
-                                    });
-
-                                    ui.horizontal(|ui| {
-                                        ui.label("        - Maximum Emission Value (optional, u64):");
-                                        let response = ui.add(
-                                            TextEdit::singleline(&mut self.inv_log_max_value_input)
-                                                .hint_text("None"),
-                                        );
-                                        if response.changed() && !self.inv_log_max_value_input.trim().is_empty() {
-                                            sanitize_u64(&mut self.inv_log_max_value_input);
-                                        }
-                                    });
-                                }
-                            }
-
-                            ui.add_space(10.0);
-
-                            // 4) Choose the distribution recipient
-                            ui.horizontal(|ui| {
-                                ui.label("     Recipient:");
-                                ComboBox::from_id_salt("perpetual_dist_recipient_selector")
-                                    .selected_text(format!("{:?}", self.perpetual_dist_recipient))
-                                    .show_ui(ui, |ui| {
-                                        ui.selectable_value(
-                                            &mut self.perpetual_dist_recipient,
-                                            TokenDistributionRecipientUI::ContractOwner,
-                                            "Contract Owner",
-                                        );
-                                        ui.selectable_value(
-                                            &mut self.perpetual_dist_recipient,
-                                            TokenDistributionRecipientUI::Identity,
-                                            "Specific Identity",
-                                        );
-                                        ui.selectable_value(
-                                            &mut self.perpetual_dist_recipient,
-                                            TokenDistributionRecipientUI::EvonodesByParticipation,
-                                            "Evonodes",
-                                        );
-                                    });
-
-                                // If user selected Identity or Group, show extra text edit
-                                match &mut self.perpetual_dist_recipient {
-                                    TokenDistributionRecipientUI::Identity => {
-                                        if self.perpetual_dist_recipient_identity_input.is_none() {
-                                            self.perpetual_dist_recipient_identity_input = Some(String::new());
-                                        }
-                                        if let Some(ref mut id) = self.perpetual_dist_recipient_identity_input {
-                                            ui.add(TextEdit::singleline(id).hint_text("Enter base58 id"));
-                                        }
-                                    }
-                                    _ => {}
-                                }
-                            });
-
-                            ui.add_space(10.0);
-
-                            ui.horizontal(|ui| {
-                                ui.label(" ");
-                                self.perpetual_distribution_rules.render_control_change_rules_ui(ui, &self.groups_ui,"Perpetual Distribution Rules", None);
-                            });
-
-                            ui.add_space(5.0);
-                        } else {
-                            self.perpetual_dist_type = PerpetualDistributionIntervalTypeUI::None;
-                        }
-
-                        ui.separator();
-
-                        // PRE-PROGRAMMED DISTRIBUTION
-                        ui.checkbox(
-                            &mut self.enable_pre_programmed_distribution,
-                            "Enable Pre-Programmed Distribution",
-                        );
-
-                        if self.enable_pre_programmed_distribution {
-                            ui.add_space(2.0);
-
-                            let mut i = 0;
-                            while i < self.pre_programmed_distributions.len() {
-                                // Clone the current entry
-                                let mut entry = self.pre_programmed_distributions[i].clone();
-
-                                // Render row
-                                ui.horizontal(|ui| {
-                                    ui.label(format!("Timestamp #{}:", i + 1));
-
-                                    // Replace text-edit timestamp with days/hours/minutes
-                                    ui.add(
-                                        egui::DragValue::new(&mut entry.days)
-                                            .prefix("Days: ")
-                                            .range(0..=3650),
-                                    );
-                                    ui.add(
-                                        egui::DragValue::new(&mut entry.hours)
-                                            .prefix("Hours: ")
-                                            .range(0..=23),
-                                    );
-                                    ui.add(
-                                        egui::DragValue::new(&mut entry.minutes)
-                                            .prefix("Minutes: ")
-                                            .range(0..=59),
-                                    );
-
-                                    ui.label("Identity:");
-                                    ui.text_edit_singleline(&mut entry.identity_str);
-
-                                    ui.label("Amount:");
-                                    ui.text_edit_singleline(&mut entry.amount_str);
-
-                                    // Remove button
-                                    if ui.button("Remove").clicked() {
-                                        self.pre_programmed_distributions.remove(i);
-                                    } else {
-                                        self.pre_programmed_distributions[i] = entry;
-                                    }
-                                });
-
-                                i += 1;
-                            }
-
-                            ui.add_space(2.0);
-
-                            // Add a button to insert a blank row
-                            ui.horizontal(|ui| {
-                                ui.label("   ");
-                                if ui.button("Add New Distribution Entry").clicked() {
-                                    self.pre_programmed_distributions.push(DistributionEntry::default());
-                                }
-                            });
-
-                            ui.add_space(2.0);
-                        }
-                    });
-
-                    ui.add_space(5.0);
-
-                    ui.collapsing("Groups", |ui| {
-                        ui.add_space(3.0);
-                        ui.label("Define one or more groups for multi-party control of the contract.");
-                        ui.add_space(2.0);
-
-                        // Add main group selection input
-                        ui.horizontal(|ui| {
-                            ui.label("Main Control Group Position:");
-                            ui.text_edit_singleline(&mut self.main_control_group_input);
-                        });
-
-                        ui.add_space(2.0);
-
-                        let mut group_to_remove = None;
-
-                        let last_group_position = self.groups_ui.len().saturating_sub(1);
-
-                        for (group_position, group_ui) in self.groups_ui.iter_mut().enumerate() {
-                            egui::collapsing_header::CollapsingState::load_with_default_open(
-                                ui.ctx(),
-                                format!("group_header_{}", group_position).into(),
-                                true,
-                            )
-                            .show_header(ui, |ui| {
-                                ui.label(format!("Group {}", group_position));
-                            })
-                            .body(|ui| {
-                                ui.add_space(3.0);
-
-                                ui.horizontal(|ui| {
-                                    ui.label("Required Power:");
-                                    ui.text_edit_singleline(&mut group_ui.required_power_str);
-                                });
-
-                                ui.label("Members:");
-                                ui.add_space(3.0);
-
-                                let mut j = 0;
-                                while j < group_ui.members.len() {
-                                    let mut member = group_ui.members[j].clone();
-
-                                    ui.horizontal(|ui| {
-                                        ui.label(format!("Member {}:", j + 1));
-
-                                        ComboBox::from_id_salt(format!("member_identity_selector_{}", j))
-                                            .width(200.0)
-                                            .selected_text(
-                                                self.identities
-                                                    .get(&Identifier::from_string(&member.identity_str, Encoding::Base58).unwrap_or_default())
-                                                    .map(|q| q.display_string())
-                                                    .unwrap_or_else(|| member.identity_str.clone()),
-                                            )
-                                            .show_ui(ui, |ui| {
-                                                for (identifier, qualified_identity) in self.identities.iter() {
-                                                    let id_str = identifier.to_string(Encoding::Base58);
-
-                                                    // Prevent duplicates unless in developer mode
-                                                    if !self.app_context.developer_mode.load(Ordering::Relaxed)
-                                                        && group_ui
-                                                        .members
-                                                        .iter()
-                                                        .enumerate().any(|(i, m)| i != j && m.identity_str == id_str)
-                                                    {
-                                                        continue;
-                                                    }
-
-                                                    if ui
-                                                        .selectable_label(false, qualified_identity.display_string())
-                                                        .clicked()
-                                                    {
-                                                        member.identity_str = id_str.clone();
-                                                    }
-                                                }
-                                            });
-
-                                        ui.text_edit_singleline(&mut member.identity_str);
-
-                                        ui.label("Power (u32):");
-                                        ui.text_edit_singleline(&mut member.power_str);
-
-                                        // Show red warning if someone else already used this identity
-                                        if self.app_context.developer_mode.load(Ordering::Relaxed)
-                                            && group_ui.members[j].identity_str != ""
-                                            && group_ui.members.iter().enumerate().any(|(i, m)| {
-                                            i > j && m.identity_str == group_ui.members[j].identity_str
-                                        })
-                                        {
-                                            ui.colored_label(
-                                                Color32::RED,
-                                                "This member is set later as part of this group (this entry will be ignored)",
-                                            );
-                                        }
-
-
-                                        if ui.button("Remove Member").clicked() {
-                                            group_ui.members.remove(j);
-                                            return; // return so we skip the assignment at the end
-                                        } else {
-                                            // Only assign back if we didn’t remove
-                                            group_ui.members[j] = member;
-                                        }
-                                    });
-
-                                    j += 1;
-                                }
-
-                                ui.add_space(3.0);
-                                if ui.button("Add Member").clicked() {
-                                    group_ui.members.push(GroupMemberUI {
-                                        identity_str: "".to_owned(),
-                                        power_str: "1".to_owned(),
-                                    });
-                                }
-
-                                ui.add_space(3.0);
-
-                                if group_position == last_group_position {
-                                    // A remove button for the entire group, only for last group
-                                    if ui.button("Remove This Group").clicked() {
-                                        group_to_remove = Some(group_position);
-                                        return;
-                                    }
-                                }
-                            });
-                        }
-
-                        if let Some(group_to_remove) = group_to_remove{
-                            self.groups_ui.remove(group_to_remove);
-                        }
-
-                        ui.add_space(5.0);
-                        if ui.button("Add New Group").clicked() {
-                            self.groups_ui.push(GroupConfigUI {
-                                required_power_str: "2".to_owned(),
-                                members: vec![GroupMemberUI {
-                                    identity_str: self.selected_identity.as_ref().map(|q| q.identity.id().to_string(Encoding::Base58)).unwrap_or_default(),
-                                    power_str: "1".to_string(),
-                                }, GroupMemberUI {
-                                    identity_str: "".to_string(),
-                                    power_str: "1".to_string(),
-                                }],
-                            });
-                        }
-                    });
-
-                    // 6) "Register Token Contract" button
-                    ui.add_space(10.0);
-                    let mut new_style = (**ui.style()).clone();
-                    new_style.spacing.button_padding = egui::vec2(10.0, 5.0);
-                    ui.set_style(new_style);
-                    ui.horizontal(|ui| {
-                        let register_button =
-                            egui::Button::new(RichText::new("Register Token Contract").color(Color32::WHITE))
-                                .fill(Color32::from_rgb(0, 128, 255))
-                                .frame(true)
-                                .corner_radius(3.0);
-                        if ui.add(register_button).clicked() {
-                            match self.parse_token_build_args() {
-                                Ok(args) => {
-                                    // If success, show the "confirmation popup"
-                                    // Or skip the popup entirely and dispatch tasks right now
-                                    self.cached_build_args = Some(args);
-                                    self.token_creator_error_message = None;
-                                    self.show_token_creator_confirmation_popup = true;
-                                },
-                                Err(err) => {
-                                    self.token_creator_error_message = Some(err);
-                                }
-                            }
-                        }
-                        let view_json_button = egui::Button::new(RichText::new("View JSON").color(Color32::WHITE))
-                            .fill(Color32::from_rgb(0, 128, 255))
-                            .frame(true)
-                            .corner_radius(3.0);
-                        if ui.add(view_json_button).clicked() {
-                            match self.parse_token_build_args() {
-                                Ok(args) => {
-                                    // We have the parsed token creation arguments
-                                    // We can now call build_data_contract_v1_with_one_token using `args`
-                                    self.cached_build_args = Some(args.clone());
-                                    let data_contract = match self.app_context.build_data_contract_v1_with_one_token(
-                                        args.identity_id,
-                                        args.token_names,
-                                        args.contract_keywords,
-                                        args.token_description,
-                                        args.should_capitalize,
-                                        args.decimals,
-                                        args.base_supply,
-                                        args.max_supply,
-                                        args.start_paused,
-                                        args.allow_transfers_to_frozen_identities,
-                                        args.keeps_history,
-                                        args.main_control_group,
-                                        args.manual_minting_rules,
-                                        args.manual_burning_rules,
-                                        args.freeze_rules,
-                                        args.unfreeze_rules,
-                                        args.destroy_frozen_funds_rules,
-                                        args.emergency_action_rules,
-                                        args.max_supply_change_rules,
-                                        args.conventions_change_rules,
-                                        args.main_control_group_change_authorized,
-                                        args.distribution_rules,
-                                        args.groups,
-                                    ) {
-                                        Ok(dc) => dc,
-                                        Err(e) => {
-                                            self.token_creator_error_message = Some(format!("Error building contract V1: {e}"));
-                                            return;
-                                        }
-                                    };
-
-                                    let data_contract_json = data_contract.to_json(self.app_context.platform_version).expect("Expected to map contract to json");
-                                    self.show_json_popup = true;
-                                    self.json_popup_text = serde_json::to_string_pretty(&data_contract_json).expect("Expected to serialize json");
-                                },
-                                Err(err_msg) => {
-                                    self.token_creator_error_message = Some(err_msg);
-                                },
-                            }
-                        }
-                    });
-                });
-        });
-
-        // 7) If the user pressed "Register Token Contract," show a popup confirmation
-        if self.show_token_creator_confirmation_popup {
-            action |= self.render_token_creator_confirmation_popup(ui);
-        }
-
-        if self.show_json_popup {
-            self.render_data_contract_json_popup(ui);
-        }
-
-        // 8) If we are waiting, show spinner / time elapsed
-        if let TokenCreatorStatus::WaitingForResult(start_time) = self.token_creator_status {
-            let now = Utc::now().timestamp() as u64;
-            let elapsed = now - start_time;
-            ui.add_space(10.0);
-            ui.horizontal(|ui| {
-                ui.label(format!(
-                    "Registering token contract... elapsed {}s",
-                    elapsed
-                ));
-                ui.add(egui::widgets::Spinner::default());
-            });
-        }
-
-        // Show an error if we have one
-        if let Some(err_msg) = &self.token_creator_error_message {
-            ui.add_space(10.0);
-            ui.colored_label(Color32::RED, format!("{err_msg}"));
-            ui.add_space(10.0);
-        }
-
-        action
-    }
-
-    /// Renders a popup window displaying the data contract JSON.
-    pub fn render_data_contract_json_popup(&mut self, ui: &mut Ui) {
-        if self.show_json_popup {
-            let mut is_open = true;
-            egui::Window::new("Data Contract JSON")
-                .collapsible(false)
-                .resizable(true)
-                .max_height(600.0)
-                .max_width(800.0)
-                .scroll(true)
-                .open(&mut is_open)
-                .show(ui.ctx(), |ui| {
-                    // Display the JSON in a multiline text box
-                    ui.add_space(4.0);
-                    ui.label("Below is the data contract JSON:");
-                    ui.add_space(4.0);
-
-                    egui::Resize::default()
-                        .id_salt("json_resize_area_for_contract")
-                        .default_size([750.0, 550.0])
-                        .max_height(ui.available_height() - 50.0)
-                        .max_width(ui.available_height() - 20.0)
-                        .show(ui, |ui| {
-                            egui::ScrollArea::vertical()
-                                .auto_shrink([false; 2])
-                                .show(ui, |ui| {
-                                    ui.monospace(&mut self.json_popup_text);
-                                });
-                        });
-
-                    ui.add_space(10.0);
-
-                    // A button to close
-                    if ui.button("Close").clicked() {
-                        self.show_json_popup = false;
-                    }
-                });
-
-            // If the user closed the window via the "x" in the corner
-            // we should reflect that in `show_json_popup`.
-            if !is_open {
-                self.show_json_popup = false;
-            }
-        }
-    }
-
     fn estimate_registration_cost(&self) -> Credits {
         let registration_fees = &self
             .app_context
@@ -4106,387 +1655,6 @@ Emits tokens in fixed amounts for specific intervals.
         fee += 200_000_000; //just an extra estimate
 
         fee
-    }
-
-    /// Shows a popup "Are you sure?" for creating the token contract
-    fn render_token_creator_confirmation_popup(&mut self, ui: &mut Ui) -> AppAction {
-        let mut action = AppAction::None;
-        let mut is_open = true;
-
-        egui::Window::new("Confirm Token Contract Registration")
-            .collapsible(false)
-            .open(&mut is_open)
-            .show(ui.ctx(), |ui| {
-                ui.label(
-                    "Are you sure you want to register a new token contract with these settings?\n",
-                );
-                let max_supply_display = if self.max_supply_input.len() == 0 {
-                    "None".to_string()
-                } else {
-                    self.max_supply_input.clone()
-                };
-                ui.label(format!(
-                    "Name: {}\nBase Supply: {}\nMax Supply: {}",
-                    self.token_names_input[0].0, self.base_supply_input, max_supply_display,
-                ));
-
-                ui.add_space(10.0);
-
-                ui.label(format!(
-                    "Estimated cost to register this token is {} Dash",
-                    self.estimate_registration_cost() as f64 / 100_000_000_000.0
-                ));
-
-                ui.add_space(10.0);
-
-                // Confirm
-                if ui.button("Confirm").clicked() {
-                    let args = match &self.cached_build_args {
-                        Some(args) => args.clone(),
-                        None => {
-                            // fallback if we didn't store them
-                            match self.parse_token_build_args() {
-                                Ok(a) => a,
-                                Err(err) => {
-                                    self.token_creator_error_message = Some(err);
-                                    self.show_token_creator_confirmation_popup = false;
-                                    action = AppAction::None;
-                                    return;
-                                }
-                            }
-                        }
-                    };
-
-                    // Now create your tasks
-                    let tasks = vec![
-                        BackendTask::TokenTask(TokenTask::RegisterTokenContract {
-                            identity: self.selected_identity.clone().unwrap(),
-                            signing_key: self.selected_key.clone().unwrap(),
-
-                            token_names: args.token_names,
-                            contract_keywords: args.contract_keywords,
-                            token_description: args.token_description,
-                            should_capitalize: args.should_capitalize,
-                            decimals: args.decimals,
-                            base_supply: args.base_supply,
-                            max_supply: args.max_supply,
-                            start_paused: args.start_paused,
-                            allow_transfers_to_frozen_identities: args
-                                .allow_transfers_to_frozen_identities,
-                            keeps_history: args.keeps_history,
-                            main_control_group: args.main_control_group,
-
-                            manual_minting_rules: args.manual_minting_rules,
-                            manual_burning_rules: args.manual_burning_rules,
-                            freeze_rules: args.freeze_rules,
-                            unfreeze_rules: args.unfreeze_rules,
-                            destroy_frozen_funds_rules: args.destroy_frozen_funds_rules,
-                            emergency_action_rules: args.emergency_action_rules,
-                            max_supply_change_rules: args.max_supply_change_rules,
-                            conventions_change_rules: args.conventions_change_rules,
-                            main_control_group_change_authorized: args
-                                .main_control_group_change_authorized,
-                            distribution_rules: args.distribution_rules,
-                            groups: args.groups,
-                        }),
-                        BackendTask::TokenTask(TokenTask::QueryMyTokenBalances),
-                    ];
-
-                    action = AppAction::BackendTasks(tasks, BackendTasksExecutionMode::Sequential);
-                    self.show_token_creator_confirmation_popup = false;
-                    let now = Utc::now().timestamp() as u64;
-                    self.token_creator_status = TokenCreatorStatus::WaitingForResult(now);
-                }
-
-                // Cancel
-                if ui.button("Cancel").clicked() {
-                    self.show_token_creator_confirmation_popup = false;
-                    action = AppAction::None;
-                }
-            });
-
-        if !is_open {
-            self.show_token_creator_confirmation_popup = false;
-        }
-
-        action
-    }
-
-    /// Once the contract creation is done (status=Complete),
-    /// render a simple "Success" screen
-    fn render_token_creator_success_screen(&mut self, ui: &mut Ui) {
-        ui.vertical_centered(|ui| {
-            ui.add_space(50.0);
-            ui.heading("Token Contract Created Successfully! 🎉");
-            ui.add_space(10.0);
-            if ui.button("Back").clicked() {
-                self.reset_token_creator();
-            }
-        });
-    }
-
-    /// Gathers user input and produces the arguments needed by
-    /// `build_data_contract_v1_with_one_token`.
-    /// Returns Err(error_msg) on invalid input.
-    pub fn parse_token_build_args(&mut self) -> Result<TokenBuildArgs, String> {
-        // 1) We must have a selected identity
-        let identity = self
-            .selected_identity
-            .clone()
-            .ok_or_else(|| "Please select an identity".to_string())?;
-        let identity_id = identity.identity.id().clone();
-
-        // Remove whitespace and parse the comma separated string into a vec
-        let mut contract_keywords = if self.contract_keywords_input.trim().is_empty() {
-            Vec::new()
-        } else {
-            self.contract_keywords_input
-                .split(',')
-                .map(|s| {
-                    let trimmed = s.trim().to_string();
-                    if trimmed.len() < 3 || trimmed.len() > 50 {
-                        Err(format!("Invalid contract keyword {}, keyword must be between 3 and 50 characters", trimmed))
-                    } else {
-                        Ok(trimmed)
-                    }
-                })
-                .collect::<Result<Vec<String>, String>>()?
-        };
-
-        // 2) Basic fields
-        if self.token_names_input.is_empty() {
-            return Err("Please enter a token name".to_string());
-        }
-        // If any name languages are duplicated, return an error
-        let mut seen_languages = HashSet::new();
-        for name_with_language in self.token_names_input.iter() {
-            if seen_languages.contains(&name_with_language.2) {
-                return Err(format!(
-                    "Duplicate token name language: {:?}",
-                    name_with_language.1
-                ));
-            }
-            seen_languages.insert(name_with_language.2);
-        }
-        let mut token_names: Vec<(String, String, String)> = Vec::new();
-        for name_with_language in self.token_names_input.iter() {
-            let language = match name_with_language.2 {
-                TokenNameLanguage::English => "en",
-                TokenNameLanguage::Arabic => "ar",
-                TokenNameLanguage::Bengali => "bn",
-                TokenNameLanguage::Burmese => "my",
-                TokenNameLanguage::Chinese => "zh",
-                TokenNameLanguage::Czech => "cs",
-                TokenNameLanguage::Dutch => "nl",
-                TokenNameLanguage::Farsi => "fa",
-                TokenNameLanguage::Filipino => "fil",
-                TokenNameLanguage::French => "fr",
-                TokenNameLanguage::German => "de",
-                TokenNameLanguage::Greek => "el",
-                TokenNameLanguage::Gujarati => "gu",
-                TokenNameLanguage::Hausa => "ha",
-                TokenNameLanguage::Hebrew => "he",
-                TokenNameLanguage::Hindi => "hi",
-                TokenNameLanguage::Hungarian => "hu",
-                TokenNameLanguage::Igbo => "ig",
-                TokenNameLanguage::Indonesian => "id",
-                TokenNameLanguage::Italian => "it",
-                TokenNameLanguage::Japanese => "ja",
-                TokenNameLanguage::Javanese => "jv",
-                TokenNameLanguage::Kannada => "kn",
-                TokenNameLanguage::Khmer => "km",
-                TokenNameLanguage::Korean => "ko",
-                TokenNameLanguage::Malay => "ms",
-                TokenNameLanguage::Malayalam => "ml",
-                TokenNameLanguage::Mandarin => "zh", // synonym for Chinese
-                TokenNameLanguage::Marathi => "mr",
-                TokenNameLanguage::Nepali => "ne",
-                TokenNameLanguage::Oriya => "or",
-                TokenNameLanguage::Pashto => "ps",
-                TokenNameLanguage::Polish => "pl",
-                TokenNameLanguage::Portuguese => "pt",
-                TokenNameLanguage::Punjabi => "pa",
-                TokenNameLanguage::Romanian => "ro",
-                TokenNameLanguage::Russian => "ru",
-                TokenNameLanguage::Serbian => "sr",
-                TokenNameLanguage::Sindhi => "sd",
-                TokenNameLanguage::Sinhala => "si",
-                TokenNameLanguage::Somali => "so",
-                TokenNameLanguage::Spanish => "es",
-                TokenNameLanguage::Swahili => "sw",
-                TokenNameLanguage::Swedish => "sv",
-                TokenNameLanguage::Tamil => "ta",
-                TokenNameLanguage::Telugu => "te",
-                TokenNameLanguage::Thai => "th",
-                TokenNameLanguage::Turkish => "tr",
-                TokenNameLanguage::Ukrainian => "uk",
-                TokenNameLanguage::Urdu => "ur",
-                TokenNameLanguage::Vietnamese => "vi",
-                TokenNameLanguage::Yoruba => "yo",
-            };
-
-            if name_with_language.0.len() < 3 || name_with_language.0.len() > 50 {
-                return Err(format!(
-                    "The name in {:?} must be between 3 and 50 characters",
-                    name_with_language.2
-                ));
-            }
-
-            if name_with_language.1.len() < 3 || name_with_language.1.len() > 50 {
-                return Err(format!(
-                    "The plural form in {:?} must be between 3 and 50 characters",
-                    name_with_language.2
-                ));
-            }
-
-            token_names.push((
-                name_with_language.0.clone(),
-                name_with_language.1.clone(),
-                language.to_owned(),
-            ));
-
-            // are we searchable?
-            if name_with_language.3 {
-                contract_keywords.push(name_with_language.0.clone());
-            }
-        }
-
-        let token_description = if self.token_description_input.len() > 0 {
-            Some(self.token_description_input.clone())
-        } else {
-            None
-        };
-        let decimals = self
-            .decimals_input
-            .parse::<u8>()
-            .map_err(|_| "Invalid decimal places amount".to_string())?;
-        let base_supply = self
-            .base_supply_input
-            .parse::<u64>()
-            .map_err(|_| "Invalid base supply amount".to_string())?;
-        let max_supply = if self.max_supply_input.is_empty() {
-            None
-        } else {
-            // If parse fails, error out
-            Some(
-                self.max_supply_input
-                    .parse::<u64>()
-                    .map_err(|_| "Invalid Max Supply".to_string())?,
-            )
-        };
-
-        let start_paused = self.start_as_paused_input;
-        let allow_transfers_to_frozen_identities = self.allow_transfers_to_frozen_identities;
-        let keeps_history = self.token_advanced_keeps_history.into();
-
-        let main_control_group = if self.main_control_group_input.is_empty() {
-            None
-        } else {
-            Some(
-                self.main_control_group_input
-                    .parse::<u16>()
-                    .map_err(|_| "Invalid main control group".to_string())?,
-            )
-        };
-
-        // 3) Convert your ActionChangeControlUI fields to real rules
-        // (or do the manual parse for each if needed)
-        let manual_minting_rules = self
-            .manual_minting_rules
-            .extract_change_control_rules("Manual Mint")?;
-        let manual_burning_rules = self
-            .manual_burning_rules
-            .extract_change_control_rules("Manual Burn")?;
-        let freeze_rules = self.freeze_rules.extract_change_control_rules("Freeze")?;
-        let unfreeze_rules = self
-            .unfreeze_rules
-            .extract_change_control_rules("Unfreeze")?;
-        let destroy_frozen_funds_rules = self
-            .destroy_frozen_funds_rules
-            .extract_change_control_rules("Destroy Frozen Funds")?;
-        let emergency_action_rules = self
-            .emergency_action_rules
-            .extract_change_control_rules("Emergency Action")?;
-        let max_supply_change_rules = self
-            .max_supply_change_rules
-            .extract_change_control_rules("Max Supply Change")?;
-        let conventions_change_rules = self
-            .conventions_change_rules
-            .extract_change_control_rules("Conventions Change")?;
-
-        // The main_control_group_change_authorized is done manually in your code,
-        // parse identity or group if needed. Reuse your existing logic:
-        let main_control_group_change_authorized =
-            self.parse_main_control_group_change_authorized()?;
-
-        // 4) Distribution data (perpetual & pre_programmed)
-        let distribution_rules = self.build_distribution_rules()?;
-
-        // 5) Groups
-        let groups = self.parse_groups()?;
-
-        // 6) Put it all in a struct
-        Ok(TokenBuildArgs {
-            identity_id,
-            token_names,
-            contract_keywords,
-            token_description,
-            should_capitalize: self.should_capitalize_input,
-            decimals,
-            base_supply,
-            max_supply,
-            start_paused,
-            allow_transfers_to_frozen_identities,
-            keeps_history,
-            main_control_group,
-
-            manual_minting_rules,
-            manual_burning_rules,
-            freeze_rules,
-            unfreeze_rules,
-            destroy_frozen_funds_rules,
-            emergency_action_rules,
-            max_supply_change_rules,
-            conventions_change_rules,
-            main_control_group_change_authorized,
-
-            distribution_rules: TokenDistributionRules::V0(distribution_rules),
-            groups,
-        })
-    }
-
-    /// Example of pulling out the logic to parse main_control_group_change_authorized
-    fn parse_main_control_group_change_authorized(
-        &mut self,
-    ) -> Result<AuthorizedActionTakers, String> {
-        match &mut self.authorized_main_control_group_change {
-            AuthorizedActionTakers::Identity(_) => {
-                if let Some(ref id_str) = self.main_control_group_change_authorized_identity {
-                    if let Ok(id) = Identifier::from_string(id_str, Encoding::Base58) {
-                        Ok(AuthorizedActionTakers::Identity(id))
-                    } else {
-                        Err("Invalid base58 identifier for main control group change authorized identity".to_owned())
-                    }
-                } else {
-                    Ok(AuthorizedActionTakers::Identity(Identifier::default()))
-                }
-            }
-            AuthorizedActionTakers::Group(_) => {
-                if let Some(ref group_str) = self.main_control_group_change_authorized_group {
-                    if let Ok(g) = group_str.parse::<u16>() {
-                        Ok(AuthorizedActionTakers::Group(g))
-                    } else {
-                        Err("Invalid group contract position for main control group".to_owned())
-                    }
-                } else {
-                    Ok(AuthorizedActionTakers::Group(0))
-                }
-            }
-            other => {
-                // For ContractOwner or NoOne, just return them as-is
-                Ok(other.clone())
-            }
-        }
     }
 
     fn build_distribution_rules(&mut self) -> Result<TokenDistributionRulesV0, String> {
@@ -4854,28 +2022,6 @@ Emits tokens in fixed amounts for specific intervals.
         Ok(map)
     }
 
-    /// Attempt to parse all group UI data into a BTreeMap<u16, Group>.
-    /// Returns an error if any row fails or duplicates a position, etc.
-    pub fn parse_groups(&self) -> Result<BTreeMap<u16, Group>, String> {
-        let mut map = BTreeMap::new();
-        for (i, g) in self.groups_ui.iter().enumerate() {
-            let (pos, group) = g
-                .parse_into_group(i as GroupContractPosition)
-                .map_err(|e| format!("Error in Group #{}: {e}", i + 1))?;
-
-            // Check for duplicates
-            if map.contains_key(&pos) {
-                return Err(format!(
-                    "Duplicate group position {pos} in Group #{}",
-                    i + 1
-                ));
-            }
-
-            map.insert(pos, group);
-        }
-        Ok(map)
-    }
-
     fn reset_token_creator(&mut self) {
         self.selected_identity = None;
         self.selected_key = None;
@@ -4981,66 +2127,6 @@ Emits tokens in fixed amounts for specific intervals.
 
         self.show_token_creator_confirmation_popup = false;
         self.token_creator_error_message = None;
-    }
-
-    fn render_no_owned_tokens(&mut self, ui: &mut Ui) -> AppAction {
-        let mut app_action = AppAction::None;
-        ui.vertical_centered(|ui| {
-            ui.add_space(20.0);
-            match self.tokens_subscreen {
-                TokensSubscreen::MyTokens => {
-                    ui.label(
-                        RichText::new("No tracked tokens.")
-                            .heading()
-                            .strong()
-                            .color(Color32::GRAY),
-                    );
-                }
-                TokensSubscreen::SearchTokens => {
-                    ui.label(
-                        RichText::new("No matching tokens found.")
-                            .heading()
-                            .strong()
-                            .color(Color32::GRAY),
-                    );
-                }
-                TokensSubscreen::TokenCreator => {
-                    ui.label(
-                        RichText::new("Cannot render token creator for some reason")
-                            .heading()
-                            .strong()
-                            .color(Color32::GRAY),
-                    );
-                }
-            }
-            ui.add_space(10.0);
-
-            ui.label("Please check back later or try refreshing the list.");
-            ui.add_space(20.0);
-            if ui.button("Refresh").clicked() {
-                if let RefreshingStatus::Refreshing(_) = self.refreshing_status {
-                    app_action = AppAction::None;
-                } else {
-                    let now = Utc::now().timestamp() as u64;
-                    self.refreshing_status = RefreshingStatus::Refreshing(now);
-                    match self.tokens_subscreen {
-                        TokensSubscreen::MyTokens => {
-                            app_action = AppAction::BackendTask(BackendTask::TokenTask(
-                                TokenTask::QueryMyTokenBalances,
-                            ));
-                        }
-                        TokensSubscreen::SearchTokens => {
-                            app_action = AppAction::Refresh;
-                        }
-                        TokensSubscreen::TokenCreator => {
-                            app_action = AppAction::Refresh;
-                        }
-                    }
-                }
-            }
-        });
-
-        app_action
     }
 
     fn add_token_to_tracked_tokens(&mut self, token_info: TokenInfo) -> Result<AppAction, String> {
@@ -5415,19 +2501,7 @@ impl ScreenLike for TokensScreen {
 
             match self.tokens_subscreen {
                 TokensSubscreen::MyTokens => {
-                    if self.all_known_tokens.is_empty() {
-                        // If no tokens, show a “no tokens found” message
-                        action |= self.render_no_owned_tokens(ui);
-                    } else {
-                        // Are we showing details for a selected token?
-                        if self.selected_token.is_some() {
-                            // Render detail view for one token
-                            action |= self.render_token_details(ui);
-                        } else {
-                            // Otherwise, show the list of all tokens
-                            self.render_token_list(ui);
-                        }
-                    }
+                    action |= self.render_my_tokens_subscreen(ui);
                 }
                 TokensSubscreen::SearchTokens => {
                     if self.selected_contract_id.is_some() {
